% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_doc.R, R/filter_na_count.R,
%   R/ttest_class.R, R/wilcox_test_class.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as_data_frame,filter_na_count-method}
\alias{as_data_frame,ttest-method}
\alias{as_data_frame,wilcox_test-method}
\title{Convert to data.frame}
\usage{
\S4method{as_data_frame}{filter_na_count}(M)

\S4method{as_data_frame}{ttest}(M)

\S4method{as_data_frame}{wilcox_test}(M)
}
\arguments{
\item{M}{a model object}
}
\value{
A data.frame of model outputs
}
\description{
convert the outputs of the input model into a data.frame.
}
\examples{
D = iris_DatasetExperiment()
M = filter_na_count(threshold=50,factor_name='Species')
M= model_apply(M,D)
df = as_data_frame(M)
}
