% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscale_class.R
\docType{class}
\name{autoscale-class}
\alias{autoscale-class}
\alias{autoscale}
\title{autoscale class}
\description{
autoscaling centres the columns of the data in a dataset object and divides by the standard deviation.
}
