% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscale_class.R
\name{autoscale}
\alias{autoscale}
\title{Autoscale}
\usage{
autoscale(mode = "data", ...)
}
\arguments{
\item{mode}{Used to control whether centring is apply to the data, the meta data or both.
Can be any one of "data","sample_meta" or "both". default is "data".}

\item{...}{additional slots and values passed to struct_class}
}
\value{
A STRUCT model object with methods for autoscaling.

struct object
}
\description{
Autoscaling centres the columns of the data in a DatasetExperiment object and divides by the standard deviation.
}
\examples{
D = iris_DatasetExperiment()
M = autoscale()
M = model_train(M,D)
M = model_predict(M,D)
}
