% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_accuracy_class.R
\docType{class}
\name{balanced_accuracy-class}
\alias{balanced_accuracy-class}
\alias{balanced_accuracy}
\title{Balanced Accuracy}
\value{
A metric object with methods for calculating balanced accuracy.
}
\description{
Balanced accuracy is the average of the true positive rate and false positive
rate.
}
\examples{
D = iris_dataset()
XCV = kfold_xval(folds=5,factor_name='Species') *
      (mean_centre() + PLSDA(number_components=2,factor_name='Species'))
MET = balanced_accuracy()
XCV = run(XCV,D,MET)

}
