% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_accuracy_class.R
\name{balanced_error}
\alias{balanced_error}
\title{Balanced error}
\usage{
balanced_error(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
balanced_error
} object. This object has no \code{output} slots.
}
\description{
Balanced Accuracy is the average proportion of correctly identified samples within each class. Balanced error is 1 - Balanced Accuracy.
}
\section{Inheritance}{

A \code{balanced_error} object inherits the following \code{struct} classes: \cr\cr
\verb{[balanced_error]} >> \verb{[metric]} >> \verb{[struct_class]}
}

\examples{
M = balanced_error()

D = iris_DatasetExperiment()
XCV = kfold_xval(folds=5,factor_name='Species') *
      (mean_centre() + PLSDA(number_components=2,factor_name='Species'))
MET = balanced_error()
XCV = run(XCV,D,MET)
}
