% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_filter_class.R
\docType{class}
\name{blank_filter.hist-class}
\alias{blank_filter.hist-class}
\alias{blank_filter.hist}
\title{plot for blank filter}
\description{
plots a histogram of the calculated fold change for the blank filter (median blank / median sample)
}
\examples{
C = blank_filter.hist()
}
