% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_class.R
\docType{class}
\name{bootstrap-class}
\alias{bootstrap-class}
\alias{bootstrap}
\title{permutation test class}
\description{
Applies a permutation test to a model or model.seq()
}
\examples{
I=bootstrap()

}
