% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_metric_class.R,
%   R/balanced_accuracy_class.R, R/calculate_doc.R, R/r_squared_class.R
\name{calculate,AUC-method}
\alias{calculate,AUC-method}
\alias{calculate,balanced_accuracy-method}
\alias{calculate,balanced_error-method}
\alias{calculate}
\alias{calculate,r_squared-method}
\title{Calculate metric}
\usage{
\S4method{calculate}{AUC}(obj, Y, Yhat)

\S4method{calculate}{balanced_accuracy}(obj, Y, Yhat)

\S4method{calculate}{balanced_error}(obj, Y, Yhat)

\S4method{calculate}{r_squared}(obj, Y, Yhat)
}
\arguments{
\item{obj}{a metric object}

\item{Y}{the true values/group labels}

\item{Yhat}{the predicted values/group labels}
}
\value{
a modified metric object
}
\description{
Calculate metric
}
\examples{
MET = metric()
calculate(MET)
}
