% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_1d_class.R
\docType{methods}
\name{chart.plot,gs_line,grid_search_1d-method}
\alias{chart.plot,gs_line,grid_search_1d-method}
\title{chart.plot}
\usage{
\S4method{chart.plot}{gs_line,grid_search_1d}(obj, dobj)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a struct object}
}
\value{
a plot object
}
\description{
Plots a chart object
}
\examples{
C = example_chart()
chart.plot(C)
}
