% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R, R/PLSDA_charts.R,
%   R/fold_change_class.R, R/PLSR_class.R, R/blank_filter_class.R,
%   R/confounders_clsq_class.R, R/dataset_chart_classes.R,
%   R/factor_barchart_class.R, R/feature_profile_class.R,
%   R/forward_selection_by_rank_class.R, R/grid_search_1d_class.R,
%   R/hca_class.R, R/kfold_xval_charts.R, R/kw_rank_sum_class.R,
%   R/mv_feature_filter_class.R, R/mv_sample_filter_class.R,
%   R/permutation_test_class.R, R/pqn_norm_method_class.R, R/rsd_filter.R,
%   R/tSNE_class.R, R/wilcox_test_class.R
\docType{methods}
\name{chart.plot,pca_correlation_plot,PCA-method}
\alias{chart.plot,pca_correlation_plot,PCA-method}
\alias{chart.plot,pca_scores_plot,PCA-method}
\alias{chart.plot,pca_biplot_plot,PCA-method}
\alias{chart.plot,pca_loadings_plot,PCA-method}
\alias{chart.plot,PCA.scree,PCA-method}
\alias{chart.plot,PCA.dstat,PCA-method}
\alias{chart.plot,plsda_scores_plot,PLSDA-method}
\alias{chart.plot,fold_change_plot,fold_change-method}
\alias{chart.plot,plsr_prediction_plot,PLSR-method}
\alias{chart.plot,plsr_residual_hist,PLSR-method}
\alias{chart.plot,plsr_qq_plot,PLSR-method}
\alias{chart.plot,plsr_cook_dist,PLSR-method}
\alias{chart.plot,blank_filter.hist,blank_filter-method}
\alias{chart.plot,confounders_lsq.barchart,confounders_clsq-method}
\alias{chart.plot,confounders_lsq.boxplot,confounders_clsq-method}
\alias{chart.plot,feature_boxplot,dataset-method}
\alias{chart.plot,mv_histogram,dataset-method}
\alias{chart.plot,mv_boxplot,dataset-method}
\alias{chart.plot,dataset.dist,dataset-method}
\alias{chart.plot,dataset.boxplot,dataset-method}
\alias{chart.plot,compare_dist,dataset-method}
\alias{chart.plot,dataset.heatmap,dataset-method}
\alias{chart.plot,dataset.factor_barchart,dataset-method}
\alias{chart.plot,feature_profile,dataset-method}
\alias{chart.plot,fs_line,forward_selection_byrank-method}
\alias{chart.plot,gs_line,grid_search_1d-method}
\alias{chart.plot,hca_dendrogram,HCA-method}
\alias{chart.plot,kfoldxcv_grid,kfold_xval-method}
\alias{chart.plot,kfoldxcv_metric,kfold_xval-method}
\alias{chart.plot,kw_p_hist,kw_rank_sum-method}
\alias{chart.plot,mv_feature_filter.hist,mv_feature_filter-method}
\alias{chart.plot,mv_sample_filter.hist,mv_sample_filter-method}
\alias{chart.plot,permutation_test.boxplot,permutation_test-method}
\alias{chart.plot,permutation_test.violin,permutation_test-method}
\alias{chart.plot,permutation_test.hist,permutation_test-method}
\alias{chart.plot,permutation_test.scatter,permutation_test-method}
\alias{chart.plot,pqn_norm.hist,pqn_norm-method}
\alias{chart.plot,rsd_filter.hist,rsd_filter-method}
\alias{chart.plot,tSNE_scatter,tSNE-method}
\alias{chart.plot,wilcox_p_hist,wilcox_test-method}
\title{chart.plot}
\usage{
\S4method{chart.plot}{pca_correlation_plot,PCA}(obj, dobj)

\S4method{chart.plot}{pca_scores_plot,PCA}(obj, dobj)

\S4method{chart.plot}{pca_biplot_plot,PCA}(obj, dobj)

\S4method{chart.plot}{pca_loadings_plot,PCA}(obj, dobj)

\S4method{chart.plot}{PCA.scree,PCA}(obj, dobj)

\S4method{chart.plot}{PCA.dstat,PCA}(obj, dobj)

\S4method{chart.plot}{plsda_scores_plot,PLSDA}(obj, dobj)

\S4method{chart.plot}{fold_change_plot,fold_change}(obj, dobj)

\S4method{chart.plot}{plsr_prediction_plot,PLSR}(obj, dobj)

\S4method{chart.plot}{plsr_residual_hist,PLSR}(obj, dobj)

\S4method{chart.plot}{plsr_qq_plot,PLSR}(obj, dobj)

\S4method{chart.plot}{plsr_cook_dist,PLSR}(obj, dobj)

\S4method{chart.plot}{blank_filter.hist,blank_filter}(obj, dobj)

\S4method{chart.plot}{confounders_lsq.barchart,confounders_clsq}(obj, dobj)

\S4method{chart.plot}{confounders_lsq.boxplot,confounders_clsq}(obj, dobj)

\S4method{chart.plot}{feature_boxplot,dataset}(obj, dobj)

\S4method{chart.plot}{mv_histogram,dataset}(obj, dobj)

\S4method{chart.plot}{mv_boxplot,dataset}(obj, dobj)

\S4method{chart.plot}{dataset.dist,dataset}(obj, dobj)

\S4method{chart.plot}{dataset.boxplot,dataset}(obj, dobj)

\S4method{chart.plot}{compare_dist,dataset}(obj, dobj, eobj)

\S4method{chart.plot}{dataset.heatmap,dataset}(obj, dobj)

\S4method{chart.plot}{dataset.factor_barchart,dataset}(obj, dobj)

\S4method{chart.plot}{feature_profile,dataset}(obj, dobj)

\S4method{chart.plot}{fs_line,forward_selection_byrank}(obj, dobj)

\S4method{chart.plot}{gs_line,grid_search_1d}(obj, dobj)

\S4method{chart.plot}{hca_dendrogram,HCA}(obj, dobj)

\S4method{chart.plot}{kfoldxcv_grid,kfold_xval}(obj, dobj)

\S4method{chart.plot}{kfoldxcv_metric,kfold_xval}(obj, dobj)

\S4method{chart.plot}{kw_p_hist,kw_rank_sum}(obj, dobj)

\S4method{chart.plot}{mv_feature_filter.hist,mv_feature_filter}(obj, dobj)

\S4method{chart.plot}{mv_sample_filter.hist,mv_sample_filter}(obj, dobj)

\S4method{chart.plot}{permutation_test.boxplot,permutation_test}(obj, dobj)

\S4method{chart.plot}{permutation_test.violin,permutation_test}(obj, dobj)

\S4method{chart.plot}{permutation_test.hist,permutation_test}(obj, dobj)

\S4method{chart.plot}{permutation_test.scatter,permutation_test}(obj, dobj)

\S4method{chart.plot}{pqn_norm.hist,pqn_norm}(obj, dobj)

\S4method{chart.plot}{rsd_filter.hist,rsd_filter}(obj, dobj)

\S4method{chart.plot}{tSNE_scatter,tSNE}(obj, dobj)

\S4method{chart.plot}{wilcox_p_hist,wilcox_test}(obj, dobj)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a struct object}
}
\value{
a plot object
}
\description{
Plots a chart object
}
\examples{
C = example_chart()
chart.plot(C)
}
