% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{compare_dist-class}
\alias{compare_dist-class}
\alias{compare_dist}
\title{Compare distributions}
\arguments{
\item{factor_name}{the sample_meta colum to use}
}
\description{
A combination of plots to compare distributions of samples/features in two
datasets
}
\examples{
D1=sbcms_dataset(filtered=FALSE)
D2=sbcms_dataset(filtered=TRUE)
C = compare_dist(factor_name='class')
chart.plot(C,D1,D2)
}
