% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders_clsq_class.R
\docType{class}
\name{confounders_clsq-class}
\alias{confounders_clsq-class}
\alias{confounders_clsq}
\title{check for confounding factors in classical least squares univariate regression}
\description{
compares the coefficients for a model without including confounding factors to models with confounding factor included.
}
\examples{
M = confounders_clsq()
}
