% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders_clsq_class.R
\name{confounders_clsq}
\alias{confounders_clsq}
\title{Check for confounding factors in ttest}
\usage{
confounders_clsq(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
A STRUCT method object with functions for applying classical least squares
}
\description{
Compares the coefficients for a ttest without including confounding factors
to models with confounding factor included. Currently only ttest is supported.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{p-value threshold for determining significance. Default alpha = 0.05.}

\item{\code{mtc}}{multiple test correction method to apply. Can be: holm, hochberg,
hommel, bonferroni, BH, BY, fdr or [none]}

\item{\code{factor_name}}{the column name of sample_meta to use in regression}

\item{\code{confounding_factors}}{the column names of factors potentially confounding
with the main factor if interest}

\item{\code{threshold}}{the threshold (between 0 and 1) for accepting a factor as confounding}
}}

\examples{
D = sbcms_DatasetExperiment()
M = filter_by_name(mode='include',dimension='variable',
        names=colnames(D$data)[1:10]) + # first 10 features
    filter_smeta(mode='exclude',levels='QC',
        factor_name='class') + # reduce to two group comparison
    confounders_clsq(factor_name = 'class',
        confounding_factors=c('sample_order','batch'))
M = model_apply(M,D)

}
