% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders_clsq_class.R
\docType{class}
\name{confounders_lsq.barchart-class}
\alias{confounders_lsq.barchart-class}
\alias{confounders_lsq.barchart}
\title{barchart of percent change}
\description{
plots a barchart of the percent change when including a confounding factor in a classical least squares model
}
\examples{
C = confounders_lsq.barchart()
}
