% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders_clsq_class.R
\name{confounders_lsq.barchart}
\alias{confounders_lsq.barchart}
\title{barchart of percent change}
\usage{
confounders_lsq.barchart(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
A STRUCT chart object

struct object
}
\description{
plots a barchart of the percent change when including a confounding factor in a classical least squares model
}
\section{Slots}{

\describe{
\item{\code{feature_to_plot}}{the name or index of the feature to be plotted}

\item{\code{threshold}}{the threshold to be plotted (in \%)}
}}

\examples{
D = sbcms_DatasetExperiment()
M = filter_by_name(mode='include',dimension='variable',
        names=colnames(D$data)[1:10]) + # first 10 features
    filter_smeta(mode='exclude',levels='QC',
        factor_name='class') + # reduce to two group comparison
    confounders_clsq(factor_name = 'class',
        confounding_factors=c('sample_order','batch'))
M = model_apply(M,D)
C = C=confounders_lsq.barchart(feature_to_plot=1,threshold=15)
chart_plot(C,M[3])

}
