% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders_clsq_class.R
\docType{class}
\name{confounders_lsq.boxplot-class}
\alias{confounders_lsq.boxplot-class}
\alias{confounders_lsq.boxplot}
\title{boxplot of percent change}
\arguments{
\item{threshold}{the threshold to be plotted (in %)}
}
\value{
A STRUCT chart object
}
\description{
Plots a boxplot of the percent change over all features when including a
confounding factor in the ttest
}
\examples{
D = sbcms_dataset()
M = filter_by_name(mode='include',dimension='variable',names=colnames(D$data)[1:10]) + # first 10 features
    filter_smeta(mode='exclude',levels='QC',factor_name='class') + # reduce to two group comparison
    confounders_clsq(factor_name = 'class',confounding_factors=c('sample_order','batch'))
M = method.apply(M,D)
C = C=confounders_lsq.box(threshold=15)
chart.plot(C,M[3])

}
