% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders_clsq_class.R
\docType{class}
\name{confounders_lsq.image-class}
\alias{confounders_lsq.image-class}
\alias{confounders_lsq.image}
\title{image of percent change}
\description{
plots an intensity image of the percent change when including a confounding factor in a classical least squares model
}
