% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounders_clsq_class.R
\name{confounders_lsq_boxplot}
\alias{confounders_lsq_boxplot}
\title{boxplot of percent change}
\usage{
confounders_lsq_boxplot(threshold = 10, ...)
}
\arguments{
\item{threshold}{the threshold to be plotted (in \%)}

\item{...}{additional slots and values passed to struct_class}
}
\value{
A STRUCT chart object

struct object
}
\description{
Plots a boxplot of the percent change over all features when including a
confounding factor in the ttest
}
\examples{
D = MTBLS79_DatasetExperiment()
M = filter_by_name(mode='include',dimension='variable',
        names=colnames(D$data)[1:10]) + # first 10 features
    filter_smeta(mode='exclude',levels='QC',
        factor_name='class') + # reduce to two group comparison
    confounders_clsq(factor_name = 'class',
        confounding_factors=c('sample_order','batch'))
M = model_apply(M,D)
C = C=confounders_lsq_boxplot(threshold=15)
chart_plot(C,M[3])

}
