% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_sum_norm_class.R
\name{constant_sum_norm}
\alias{constant_sum_norm}
\title{Normalisation to constant sum}
\usage{
constant_sum_norm(scaling_factor = 1, ...)
}
\arguments{
\item{scaling_factor}{The sum of all values for a sample after normalisation. Default = 1.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Applies normalisation to a constant sum, such that the sum of values for each sample
after normalisation is equal to 1 (or a scaling factor, if specified).
}
\examples{
M = constant_sum_norm()

}
