% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_coef_class.R
\docType{class}
\name{corr_coef-class}
\alias{corr_coef-class}
\alias{corr_coef}
\title{correlation model class}
\description{
correlation model class. Calculate correlation between features and continuous variables
}
\examples{
M = corr_coef()

}
