% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{dataset.dist-class}
\alias{dataset.dist-class}
\alias{dataset.dist}
\title{Distribution plot}
\arguments{
\item{factor_name}{the sample_meta column to use}

\item{per_class}{= [TRUE] or FALSE to plot distrubutions for each level in
factor_name}
}
\description{
Visualise distributions of values in features/samples.
}
\examples{
D = sbcms_dataset()
C = dataset.dist()
chart.plot(C,D)

}
