% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_barchart_class.R
\docType{class}
\name{dataset.factor_barchart-class}
\alias{dataset.factor_barchart-class}
\alias{dataset.factor_barchart}
\title{dataset.factor_barchart chart class}
\description{
Bar charts based on groupings by factor.
}
