% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{dataset.heatmap-class}
\alias{dataset.heatmap-class}
\alias{dataset.heatmap}
\title{dataset.heatmap class}
\description{
plots a dataset as a heatmap
}
\examples{
C = dataset.heatmap()
}
