% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{DatasetExperiment.dist}
\alias{DatasetExperiment.dist}
\title{Distribution plot}
\usage{
DatasetExperiment.dist(...)
}
\arguments{
\item{factor_name}{the sample_meta column to use}

\item{per_class}{= [TRUE] or FALSE to plot distrubutions for each level in
factor_name}
}
\description{
Visualise distributions of values in features/samples.
}
\examples{
D = sbcms_DatasetExperiment()
C = DatasetExperiment.dist(factor_name='class')
chart_plot(C,D)

}
