% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{DatasetExperiment.heatmap}
\alias{DatasetExperiment.heatmap}
\title{DatasetExperiment.heatmap class}
\usage{
DatasetExperiment.heatmap(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
plots a DatasetExperiment as a heatmap
}
\examples{
C = DatasetExperiment.heatmap()
}
