% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{DatasetExperiment_boxplot}
\alias{DatasetExperiment_boxplot}
\title{DatasetExperiment boxplot}
\usage{
DatasetExperiment_boxplot(
  factor_name,
  by_sample = TRUE,
  per_class = TRUE,
  number = 50,
  ...
)
}
\arguments{
\item{factor_name}{the column name of sample_meta to use}

\item{by_sample}{[TRUE] or FALSE to plot by samples or features respectively}

\item{per_class}{[TRUE] or FALSE to plot per level in factro_name}

\item{number}{the number of samples/features to plot}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Boxplot of values per sample/feature in a DatasetExperiment
}
\examples{
D = MTBLS79_DatasetExperiment()
C = DatasetExperiment_boxplot(factor_name='class',number=10,per_class=FALSE)
chart_plot(C,D)

}
