% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{DatasetExperiment_dist}
\alias{DatasetExperiment_dist}
\title{Distribution plot}
\usage{
DatasetExperiment_dist(factor_name, per_class = TRUE, ...)
}
\arguments{
\item{factor_name}{the sample_meta column to use}

\item{per_class}{= [TRUE] or FALSE to plot distrubutions for each level in
factor_name}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Visualise distributions of values in features/samples.
}
\examples{
D = MTBLS79_DatasetExperiment()
C = DatasetExperiment_dist(factor_name='class')
chart_plot(C,D)

}
