% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_barchart_class.R
\name{DatasetExperiment_factor_barchart}
\alias{DatasetExperiment_factor_barchart}
\title{DatasetExperiment_factor_barchart class}
\usage{
DatasetExperiment_factor_barchart(feature_to_plot, factor_names, ...)
}
\arguments{
\item{feature_to_plot}{Column ID of feature to plot.}

\item{factor_names}{Names(s) of factors to plot for a feature}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Bar charts based on groupings by factor. Can plot up to three factors.
}
\examples{
D = iris_DatasetExperiment()
C = DatasetExperiment_factor_barchart(factor_names='Species',feature_to_plot='Petal.Width')
chart_plot(C,D)

}
