% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ratio_filter_class.R
\docType{class}
\name{dratio_filter-class}
\alias{dratio_filter-class}
\alias{dratio_filter}
\title{d_ratio filter}
\description{
filters features based on the d_ratio of technical variance vs sample variance
}
