% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{feature_boxplot}
\alias{feature_boxplot}
\title{Feature boxplot}
\usage{
feature_boxplot(
  label_outliers = TRUE,
  feature_to_plot,
  factor_name,
  show_counts = TRUE,
  ...
)
}
\arguments{
\item{label_outliers}{(logical) Label outliers. Allowed values are limited to the following: \itemize{\item{\code{"TRUE"}: The index for outlier samples is included on the plot.}\item{\code{"FALSE"}: No labels are displayed.}} The default is \code{TRUE}.}

\item{feature_to_plot}{(character, numeric, integer) The column name of the plotted feature.}

\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{show_counts}{(logical) Show counts. Allowed values are limited to the following: \itemize{\item{\code{"TRUE"}: The number of samples for each box is displayed.}\item{\code{"FALSE"}: The number of samples for each box is not displayed.}} The default is \code{TRUE}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{feature_boxplot} object.
}
\description{
A boxplot to visualise the distribution of values within a feature.
}
\examples{
D = MTBLS79_DatasetExperiment
C = feature_boxplot(factor_name='Species',feature_to_plot='Petal.Width')
chart_plot(C,D)
}
