% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_profile_class.R
\docType{class}
\name{feature_profile-class}
\alias{feature_profile-class}
\alias{feature_profile}
\title{feature_profile class}
\description{
plots a feature vs run order
}
\examples{
C = feature_profile()
}
