% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_profile_class.R
\name{feature_profile}
\alias{feature_profile}
\title{Feature profile class}
\usage{
feature_profile(
  run_order,
  qc_label,
  qc_column,
  colour_by,
  feature_to_plot,
  ...
)
}
\arguments{
\item{run_order}{the sample_meta column containing the measurement
order of the samples}

\item{qc_label}{the label used to identify QC samples}

\item{qc_column}{the sample_meta column containing the QC labels}

\item{colour_by}{the sample_meta column to use to colour the plot}

\item{feature_to_plot}{the column id of the feature to plot}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Scatter plot of a feature against measurement order with limits for samples
and quality control samples.
}
\examples{
D = MTBLS79_DatasetExperiment()
C = feature_profile(run_order='sample_order',
    qc_label='QC',
    qc_column='class',
    colour_by='class',
    feature_to_plot=1)
chart_plot(C,D)

}
