% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{feature_scatter-class}
\alias{feature_scatter-class}
\alias{feature_scatter}
\title{feature_scatter class}
\description{
plots a scatterplot of a chosen feature against a chosen factor
}
