% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_name_class.R
\docType{class}
\name{filter_by_name-class}
\alias{filter_by_name-class}
\alias{filter_by_name}
\title{filter by name}
\description{
a filter to reduce a dataset object based on row or column labels.
}
\examples{
M = filter_by_name()
}
