% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_na_count.R
\docType{class}
\name{filter_na_count-class}
\alias{filter_na_count-class}
\alias{filter_na_count}
\title{filter features by number of NA per class}
\description{
filters features by the number of NA per class
}
