% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_smeta_class.R
\docType{class}
\name{filter_smeta-class}
\alias{filter_smeta-class}
\alias{filter_smeta}
\title{filter by sample meta class}
\description{
a Filter to reduce a dataset object based on sample meta data labels.
}
