% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_exact_class.R
\name{fisher_exact}
\alias{fisher_exact}
\title{fisher_exact class}
\usage{
fisher_exact(alpha = 0.05, mtc = "fdr", factor_name, factor_pred, ...)
}
\arguments{
\item{alpha}{the p-value threshold to declare a result 'significant'}

\item{mtc}{multiple test correction method}

\item{factor_name}{the sample_meta column to use}

\item{factor_pred}{A data.frame, with a factor of predicted group labels to
compare with factor_name. Can be a data frame with a factor of predictions
for each feature.'}

\item{...}{additional slots and values passed to struct_class}
}
\value{
A struct model with functions for applying fisher exact test.
}
\description{
Fisher's exact test (FET). Applies FET for all features in a DatasetExperiment.
}
\examples{
# load some data
D=sbcms_DatasetExperiment()

# prepare predictions based on NA
pred=as.data.frame(is.na(D$data))
pred=lapply(pred,factor,levels=c(TRUE,FALSE))
pred=as.data.frame(pred)

# apply method
M = fisher_exact(alpha=0.05,mtc='fdr',factor_name='class',factor_pred=pred)
M=model_apply(M,D)

}
