% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_class.R
\name{fold_change}
\alias{fold_change}
\title{fold change class}
\usage{
fold_change(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
Calculates fold change between groups for all features in a DatasetExperiment, based on
a log transform and t-test.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{confidence level to use for intervals}

\item{\code{factor_name}}{the sample_meta column to use}

\item{\code{paired}}{TRUE or [FALSE] to account for paired samples}

\item{\code{sample_name}}{the sample_meta column name to use for a paired samples}

\item{\code{threshold}}{a threshold to define fold change as 'significant'.}

\item{\code{control_group}}{a level of factor name to use as the control group for
calculations.}
}}

\examples{
D = sbcms_DatasetExperiment()
M = fold_change(factor_name='class')
M = model_apply(M,D)

}
