% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_int_class.R
\docType{class}
\name{fold_change_int-class}
\alias{fold_change_int-class}
\alias{fold_change_int}
\title{fold change for interactions class}
\description{
calculates fold change between groups for all factors and interactions
}
\examples{
M = fold_change_int()
}
