% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_class.R
\name{fold_change_plot}
\alias{fold_change_plot}
\title{fold_change plot}
\usage{
fold_change_plot(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
plots fold change
}
\examples{
C = fold_change_plot()
}
