% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_selection_by_rank_class.R
\docType{class}
\name{forward_selection_byrank-class}
\alias{forward_selection_byrank-class}
\alias{forward_selection_byrank}
\title{forward selection by rank}
\description{
--insert class description here--
}
