% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_selection_by_rank_class.R
\docType{class}
\name{fs_line-class}
\alias{fs_line-class}
\alias{fs_line}
\title{forward_selection_plot}
\description{
Plots the result of the evaluated models against the values the number of
features within the search range for forward_selection_by_rank objects.
}
\examples{
# some data
D = sbcms_dataset(filtered=TRUE)

# normalise, impute and scale then remove QCs
P = pqn_norm(qc_label='QC',factor_name='class') +
    knn_impute(neighbours=5) +
    glog_transform(qc_label='QC',factor_name='class') +
    filter_smeta(mode='exclude',levels='QC',factor_name='class')
P = model.apply(P,D)
D = predicted(P)

# forward selection using a PLSDA model
M = forward_selection_byrank(factor_name='class',
                             min_no_vars=2,
                             max_no_vars=11,
                             variable_rank=1:2063) *
    (mean_centre() + PLSDA(number_components=1,
                           factor_name='class'))
M = run(M,D,balanced_accuracy())

# chart
C = fs_line()
chart.plot(C,M)

}
