% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_class.R
\docType{class}
\name{glog_transform-class}
\alias{glog_transform-class}
\alias{glog_transform}
\title{glog transform}
\description{
applies a glog transform to the input data
}
\examples{
M = glog_transform()
}
