% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_1d_class.R
\docType{class}
\name{gs_line-class}
\alias{gs_line-class}
\alias{gs_line}
\title{grid_search_plot}
\description{
plots the result of the evaluated models for against the values of the optimisation paramter within the search range.
}
\examples{
C = gs_line()
}
