% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_class.R
\name{hca_dendrogram}
\alias{hca_dendrogram}
\title{HCA dendrogram}
\usage{
hca_dendrogram(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
hca_dendrogram
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A dendrogram visualising the clustering by HCA.
}
\details{
This object makes use of functionality from the following packages:\itemize{  \item{\code{ggdendro}}}
}
\section{Inheritance}{

A \code{hca_dendrogram} object inherits the following \code{struct} classes: \cr\cr
\verb{[hca_dendrogram]} >> \verb{[chart]} >> \verb{[struct_class]}
}

\examples{
M = hca_dendrogram()

C = hca_dendrogram()
}
\references{
de Vries A, Ripley BD (2024). \emph{ggdendro: Create Dendrograms and Tree
Diagrams Using 'ggplot2'}. doi:10.32614/CRAN.package.ggdendro
\url{https://doi.org/10.32614/CRAN.package.ggdendro}, R package version
0.2.0, \url{https://CRAN.R-project.org/package=ggdendro}.
}
