% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_xval2_class.R
\docType{class}
\name{kfold_xval2-class}
\alias{kfold_xval2-class}
\alias{kfold_xval2}
\title{kfold_xval model class}
\description{
Applies k-fold crossvalidation to a model or model.seq()
}
\examples{
I = kfold_xval2()
}
