% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_xval_charts.R
\name{kfoldxcv_grid}
\alias{kfoldxcv_grid}
\title{kfoldxcv_grid class}
\usage{
kfoldxcv_grid(factor_name, ...)
}
\arguments{
\item{factor_name}{The sample_meta column name to use.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plot of cross validation predictions vs the true values.
}
\examples{
D = iris_DatasetExperiment()
I = kfold_xval(factor_name='Species') *
    (mean_centre() + PLSDA(factor_name='Species'))
I = run(I,D,balanced_accuracy())

C = kfoldxcv_grid(factor_name='Species')
chart_plot(C,I)

}
