% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_xval_charts.R
\docType{class}
\name{kfoldxcv_metric-class}
\alias{kfoldxcv_metric-class}
\alias{kfoldxcv_metric}
\title{kfoldxcv_metric class}
\description{
box plot of cross validation results
}
\examples{
C = kfoldxcv_metric()
}
