% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_rank_sum_class.R
\docType{class}
\name{kw_rank_sum-class}
\alias{kw_rank_sum-class}
\alias{kw_rank_sum}
\title{kruskal-wallis model class}
\description{
kruskal-wallis model class. Calculate kw-test for all features in a dataset.
A non-parametric 1-way ANOVA.
}
