% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_class.R
\docType{class}
\name{linear_model-class}
\alias{linear_model-class}
\alias{linear_model}
\title{linear model class}
\description{
wrapper for R lm. Intended to be inherited by other classes e.g. for ANOVA but can be used as is
}
