% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{log transform}
\usage{
log_transform(base = 10, ...)
}
\arguments{
\item{base}{The base of the logarithm. Default is 10, resulting in a log10 transformation of the data.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Applies a log transform to the input data
}
\examples{
M = log_transform()
}
