% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_centre_class.R
\docType{class}
\name{mean_centre-class}
\alias{mean_centre-class}
\alias{mean_centre}
\title{mean_centre model class}
\description{
Mean centres the columns a dataset object
}
\examples{
M = mean_centre()
}
