% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_medians.R
\name{mean_of_medians}
\alias{mean_of_medians}
\title{Mean of medians}
\usage{
mean_of_medians(factor_name, ...)
}
\arguments{
\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{mean_of_medians} object with the following \code{output} slots:
\tabular{ll}{
\code{transformed} \tab          (DatasetExperiment) Data after the tranformation has been applied. \cr
}
}
\description{
The data matrix is normalised by the mean of the median of each factor level.
}
\section{Inheritance}{

A \code{mean_of_medians} object inherits the following \code{struct} classes: \cr\cr
\verb{[mean_of_medians]} >> \verb{[model]} >> \verb{[struct_class]}
}

\examples{
M = mean_of_medians(
      factor_name = "V1")

D = iris_DatasetExperiment()
M = mean_of_medians(factor_name='Species')
M = model_apply(M,D)
}
