% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_name_class.R
\docType{methods}
\name{method.apply,filter_by_name,dataset-method}
\alias{method.apply,filter_by_name,dataset-method}
\title{Apply method}
\usage{
\S4method{method.apply}{filter_by_name,dataset}(M, D)
}
\arguments{
\item{M}{a method object}

\item{D}{another object used by the first}
}
\value{
Returns a modified method object
}
\description{
Applies method to the input dataset
}
