% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effect_class.R
\docType{class}
\name{mixed_effect-class}
\alias{mixed_effect-class}
\alias{mixed_effect}
\title{Mixed Effects model class}
\description{
Mixed Effects model class. Applies RE model for all features in a dataset
}
