% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_class.R, R/HSD_class.R,
%   R/mixed_effect_class.R, R/HSDEM_class.R, R/fold_change_class.R,
%   R/PLSFC_class.R, R/blank_filter_class.R, R/classical_lsq_class.R,
%   R/confounders_clsq_class.R, R/corr_coef_class.R, R/d_ratio_filter_class.R,
%   R/filter_by_name_class.R, R/filter_na_count.R, R/filter_smeta_class.R,
%   R/fisher_exact_class.R, R/fold_change_int_class.R, R/glog_class.R,
%   R/hca_class.R, R/knn_impute_class.R, R/kw_rank_sum_class.R,
%   R/log_transform.R, R/model_apply_doc.R, R/mv_feature_filter_class.R,
%   R/mv_sample_filter_class.R, R/pairs_filter_class.R,
%   R/pqn_norm_method_class.R, R/prop_na_class.R, R/rsd_filter.R, R/sb_corr.R,
%   R/split_data_class.R, R/tSNE_class.R, R/ttest_class.R, R/vec_norm_class.R,
%   R/wilcox_test_class.R
\docType{methods}
\name{model.apply,ANOVA,dataset-method}
\alias{model.apply,ANOVA,dataset-method}
\alias{model.apply,HSD,dataset-method}
\alias{model.apply,mixed_effect,dataset-method}
\alias{model.apply,HSDEM,dataset-method}
\alias{model.apply,fold_change,dataset-method}
\alias{model.apply,PLSFC,dataset-method}
\alias{model.apply,blank_filter,dataset-method}
\alias{model.apply,classical_lsq,dataset-method}
\alias{model.apply,confounders_clsq,dataset-method}
\alias{model.apply,corr_coef,dataset-method}
\alias{model.apply,dratio_filter,dataset-method}
\alias{model.apply,filter_by_name,dataset-method}
\alias{model.apply,filter_na_count,dataset-method}
\alias{model.apply,filter_smeta,dataset-method}
\alias{model.apply,fisher_exact,dataset-method}
\alias{model.apply,fold_change_int,dataset-method}
\alias{model.apply,glog_transform,dataset-method}
\alias{model.apply,HCA,dataset-method}
\alias{model.apply,knn_impute,dataset-method}
\alias{model.apply,kw_rank_sum,dataset-method}
\alias{model.apply,log_transform,dataset-method}
\alias{model.apply}
\alias{model.apply,mv_feature_filter,dataset-method}
\alias{model.apply,mv_sample_filter,dataset-method}
\alias{model.apply,pairs_filter,dataset-method}
\alias{model.apply,pqn_norm,dataset-method}
\alias{model.apply,prop_na,dataset-method}
\alias{model.apply,rsd_filter,dataset-method}
\alias{model.apply,sb_corr,dataset-method}
\alias{model.apply,split_data,dataset-method}
\alias{model.apply,tSNE,dataset-method}
\alias{model.apply,ttest,dataset-method}
\alias{model.apply,vec_norm,dataset-method}
\alias{model.apply,wilcox_test,dataset-method}
\title{Apply method}
\usage{
\S4method{model.apply}{ANOVA,dataset}(M, D)

\S4method{model.apply}{HSD,dataset}(M, D)

\S4method{model.apply}{mixed_effect,dataset}(M, D)

\S4method{model.apply}{HSDEM,dataset}(M, D)

\S4method{model.apply}{fold_change,dataset}(M, D)

\S4method{model.apply}{PLSFC,dataset}(M, D)

\S4method{model.apply}{blank_filter,dataset}(M, D)

\S4method{model.apply}{classical_lsq,dataset}(M, D)

\S4method{model.apply}{confounders_clsq,dataset}(M, D)

\S4method{model.apply}{corr_coef,dataset}(M, D)

\S4method{model.apply}{dratio_filter,dataset}(M, D)

\S4method{model.apply}{filter_by_name,dataset}(M, D)

\S4method{model.apply}{filter_na_count,dataset}(M, D)

\S4method{model.apply}{filter_smeta,dataset}(M, D)

\S4method{model.apply}{fisher_exact,dataset}(M, D)

\S4method{model.apply}{fold_change_int,dataset}(M, D)

\S4method{model.apply}{glog_transform,dataset}(M, D)

\S4method{model.apply}{HCA,dataset}(M, D)

\S4method{model.apply}{knn_impute,dataset}(M, D)

\S4method{model.apply}{kw_rank_sum,dataset}(M, D)

\S4method{model.apply}{log_transform,dataset}(M, D)

\S4method{model.apply}{mv_feature_filter,dataset}(M, D)

\S4method{model.apply}{mv_sample_filter,dataset}(M, D)

\S4method{model.apply}{pairs_filter,dataset}(M, D)

\S4method{model.apply}{pqn_norm,dataset}(M, D)

\S4method{model.apply}{prop_na,dataset}(M, D)

\S4method{model.apply}{rsd_filter,dataset}(M, D)

\S4method{model.apply}{sb_corr,dataset}(M, D)

\S4method{model.apply}{split_data,dataset}(M, D)

\S4method{model.apply}{tSNE,dataset}(M, D)

\S4method{model.apply}{ttest,dataset}(M, D)

\S4method{model.apply}{vec_norm,dataset}(M, D)

\S4method{model.apply}{wilcox_test,dataset}(M, D)
}
\arguments{
\item{M}{a method object}

\item{D}{another object used by the first}
}
\value{
Returns a modified method object
}
\description{
Applies method to the input dataset
}
\examples{
M=model()
model.apply(M,dataset())
}
