% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_class.R, R/PLSDA_class.R, R/PLSR_class.R,
%   R/autoscale_class.R, R/linear_model_class.R, R/mean_centre_class.R,
%   R/model_predict_doc.R
\docType{methods}
\name{model.predict,PCA,dataset-method}
\alias{model.predict,PCA,dataset-method}
\alias{model.predict,PLSDA,dataset-method}
\alias{model.predict,PLSR,dataset-method}
\alias{model.predict,autoscale,dataset-method}
\alias{model.predict,linear_model,dataset-method}
\alias{model.predict,mean_centre,dataset-method}
\alias{model.predict}
\title{Model prediction}
\usage{
\S4method{model.predict}{PCA,dataset}(M, D)

\S4method{model.predict}{PLSDA,dataset}(M, D)

\S4method{model.predict}{PLSR,dataset}(M, D)

\S4method{model.predict}{autoscale,dataset}(M, D)

\S4method{model.predict}{linear_model,dataset}(M, D)

\S4method{model.predict}{mean_centre,dataset}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified model object
}
\description{
Apply a model using the input dataset. Assumes the model is trained
first.
}
\examples{
M = example_model()
M = model.predict(M,iris_dataset())
}
