% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_centre_class.R, R/model_reverse_doc.R
\docType{methods}
\name{model.reverse,mean_centre,dataset-method}
\alias{model.reverse,mean_centre,dataset-method}
\alias{model.reverse}
\title{Reverse preprocessing}
\usage{
\S4method{model.reverse}{mean_centre,dataset}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified dataset object
}
\description{
Reverse the effect of a preprocessing step on a dataset.
}
\examples{
M = example_model()
D = model.reverse(M,iris_dataset())
}
