% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_class.R, R/PLSDA_class.R, R/PLSR_class.R,
%   R/autoscale_class.R, R/blank_filter_class.R, R/d_ratio_filter_class.R,
%   R/filter_by_name_class.R, R/filter_na_count.R, R/filter_smeta_class.R,
%   R/linear_model_class.R, R/mean_centre_class.R, R/model_train_doc.R,
%   R/mv_feature_filter_class.R, R/mv_sample_filter_class.R,
%   R/pairs_filter_class.R, R/vec_norm_class.R
\docType{methods}
\name{model.train,PCA,dataset-method}
\alias{model.train,PCA,dataset-method}
\alias{model.train,PLSDA,dataset-method}
\alias{model.train,PLSR,dataset-method}
\alias{model.train,autoscale,dataset-method}
\alias{model.train,blank_filter,dataset-method}
\alias{model.train,dratio_filter,dataset-method}
\alias{model.train,filter_by_name,dataset-method}
\alias{model.train,filter_na_count,dataset-method}
\alias{model.train,filter_smeta,dataset-method}
\alias{model.train,linear_model,dataset-method}
\alias{model.train,mean_centre,dataset-method}
\alias{model.train}
\alias{model.train,mv_feature_filter,dataset-method}
\alias{model.train,mv_sample_filter,dataset-method}
\alias{model.train,pairs_filter,dataset-method}
\alias{model.train,vec_norm,dataset-method}
\title{Train a model}
\usage{
\S4method{model.train}{PCA,dataset}(M, D)

\S4method{model.train}{PLSDA,dataset}(M, D)

\S4method{model.train}{PLSR,dataset}(M, D)

\S4method{model.train}{autoscale,dataset}(M, D)

\S4method{model.train}{blank_filter,dataset}(M, D)

\S4method{model.train}{dratio_filter,dataset}(M, D)

\S4method{model.train}{filter_by_name,dataset}(M, D)

\S4method{model.train}{filter_na_count,dataset}(M, D)

\S4method{model.train}{filter_smeta,dataset}(M, D)

\S4method{model.train}{linear_model,dataset}(M, D)

\S4method{model.train}{mean_centre,dataset}(M, D)

\S4method{model.train}{mv_feature_filter,dataset}(M, D)

\S4method{model.train}{mv_sample_filter,dataset}(M, D)

\S4method{model.train}{pairs_filter,dataset}(M, D)

\S4method{model.train}{vec_norm,dataset}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified model object
}
\description{
Trains a model using the input dataset
}
\examples{
M = example_model()
M = model.train(M,iris_dataset())
}
