% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_class.R, R/HSD_class.R,
%   R/mixed_effect_class.R, R/HSDEM_class.R, R/classical_lsq_class.R,
%   R/confounders_clsq_class.R, R/constant_sum_norm_class.R,
%   R/corr_coef_class.R, R/filter_smeta_class.R, R/fisher_exact_class.R,
%   R/fold_change_class.R, R/fold_change_int_class.R, R/hca_class.R,
%   R/knn_impute_class.R, R/kw_rank_sum_class.R, R/log_transform.R,
%   R/mean_of_medians.R, R/model_apply_doc.R, R/nroot_transform_class.R,
%   R/pairs_filter_class.R, R/prop_na_class.R, R/rsd_filter.R, R/sb_corr.R,
%   R/split_data_class.R, R/stratified_split_class.R, R/tSNE_class.R,
%   R/ttest_class.R, R/vec_norm_class.R, R/wilcox_test_class.R
\name{model_apply,ANOVA,DatasetExperiment-method}
\alias{model_apply,ANOVA,DatasetExperiment-method}
\alias{model_apply,HSD,DatasetExperiment-method}
\alias{model_apply,mixed_effect,DatasetExperiment-method}
\alias{model_apply,HSDEM,DatasetExperiment-method}
\alias{model_apply,classical_lsq,DatasetExperiment-method}
\alias{model_apply,confounders_clsq,DatasetExperiment-method}
\alias{model_apply,constant_sum_norm,DatasetExperiment-method}
\alias{model_apply,corr_coef,DatasetExperiment-method}
\alias{model_apply,filter_smeta,DatasetExperiment-method}
\alias{model_apply,fisher_exact,DatasetExperiment-method}
\alias{model_apply,fold_change,DatasetExperiment-method}
\alias{model_apply,fold_change_int,DatasetExperiment-method}
\alias{model_apply,HCA,DatasetExperiment-method}
\alias{model_apply,knn_impute,DatasetExperiment-method}
\alias{model_apply,kw_rank_sum,DatasetExperiment-method}
\alias{model_apply,log_transform,DatasetExperiment-method}
\alias{model_apply,mean_of_medians,DatasetExperiment-method}
\alias{model_apply}
\alias{model_apply,nroot_transform,DatasetExperiment-method}
\alias{model_apply,pairs_filter,DatasetExperiment-method}
\alias{model_apply,prop_na,DatasetExperiment-method}
\alias{model_apply,rsd_filter,DatasetExperiment-method}
\alias{model_apply,sb_corr,DatasetExperiment-method}
\alias{model_apply,split_data,DatasetExperiment-method}
\alias{model_apply,stratified_split,DatasetExperiment-method}
\alias{model_apply,tSNE,DatasetExperiment-method}
\alias{model_apply,ttest,DatasetExperiment-method}
\alias{model_apply,vec_norm,DatasetExperiment-method}
\alias{model_apply,wilcox_test,DatasetExperiment-method}
\title{Apply method}
\usage{
\S4method{model_apply}{ANOVA,DatasetExperiment}(M, D)

\S4method{model_apply}{HSD,DatasetExperiment}(M, D)

\S4method{model_apply}{mixed_effect,DatasetExperiment}(M, D)

\S4method{model_apply}{HSDEM,DatasetExperiment}(M, D)

\S4method{model_apply}{classical_lsq,DatasetExperiment}(M, D)

\S4method{model_apply}{confounders_clsq,DatasetExperiment}(M, D)

\S4method{model_apply}{constant_sum_norm,DatasetExperiment}(M, D)

\S4method{model_apply}{corr_coef,DatasetExperiment}(M, D)

\S4method{model_apply}{filter_smeta,DatasetExperiment}(M, D)

\S4method{model_apply}{fisher_exact,DatasetExperiment}(M, D)

\S4method{model_apply}{fold_change,DatasetExperiment}(M, D)

\S4method{model_apply}{fold_change_int,DatasetExperiment}(M, D)

\S4method{model_apply}{HCA,DatasetExperiment}(M, D)

\S4method{model_apply}{knn_impute,DatasetExperiment}(M, D)

\S4method{model_apply}{kw_rank_sum,DatasetExperiment}(M, D)

\S4method{model_apply}{log_transform,DatasetExperiment}(M, D)

\S4method{model_apply}{mean_of_medians,DatasetExperiment}(M, D)

\S4method{model_apply}{nroot_transform,DatasetExperiment}(M, D)

\S4method{model_apply}{pairs_filter,DatasetExperiment}(M, D)

\S4method{model_apply}{prop_na,DatasetExperiment}(M, D)

\S4method{model_apply}{rsd_filter,DatasetExperiment}(M, D)

\S4method{model_apply}{sb_corr,DatasetExperiment}(M, D)

\S4method{model_apply}{split_data,DatasetExperiment}(M, D)

\S4method{model_apply}{stratified_split,DatasetExperiment}(M, D)

\S4method{model_apply}{tSNE,DatasetExperiment}(M, D)

\S4method{model_apply}{ttest,DatasetExperiment}(M, D)

\S4method{model_apply}{vec_norm,DatasetExperiment}(M, D)

\S4method{model_apply}{wilcox_test,DatasetExperiment}(M, D)
}
\arguments{
\item{M}{a method object}

\item{D}{another object used by the first}
}
\value{
Returns a modified method object
}
\description{
Applies method to the input DatasetExperiment
}
\examples{
M=model()
model_apply(M,DatasetExperiment())
}
