% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_class.R, R/PCA_class.R, R/PLSR_class.R,
%   R/PLSDA_class.R, R/autoscale_class.R, R/blank_filter_class.R,
%   R/constant_sum_norm_class.R, R/d_ratio_filter_class.R,
%   R/filter_by_name_class.R, R/filter_na_count.R, R/filter_smeta_class.R,
%   R/glog_class.R, R/linear_model_class.R, R/mean_centre_class.R,
%   R/model_predict_doc.R, R/mv_feature_filter_class.R,
%   R/mv_sample_filter_class.R, R/paretoscale_class.R,
%   R/pqn_norm_method_class.R, R/svm_classifier_class.R, R/vec_norm_class.R
\name{model_predict,DFA,DatasetExperiment-method}
\alias{model_predict,DFA,DatasetExperiment-method}
\alias{model_predict,PCA,DatasetExperiment-method}
\alias{model_predict,PLSR,DatasetExperiment-method}
\alias{model_predict,PLSDA,DatasetExperiment-method}
\alias{model_predict,autoscale,DatasetExperiment-method}
\alias{model_predict,blank_filter,DatasetExperiment-method}
\alias{model_predict,constant_sum_norm,DatasetExperiment-method}
\alias{model_predict,dratio_filter,DatasetExperiment-method}
\alias{model_predict,filter_by_name,DatasetExperiment-method}
\alias{model_predict,filter_na_count,DatasetExperiment-method}
\alias{model_predict,filter_smeta,DatasetExperiment-method}
\alias{model_predict,glog_transform,DatasetExperiment-method}
\alias{model_predict,linear_model,DatasetExperiment-method}
\alias{model_predict,mean_centre,DatasetExperiment-method}
\alias{model_predict}
\alias{model_predict,mv_feature_filter,DatasetExperiment-method}
\alias{model_predict,mv_sample_filter,DatasetExperiment-method}
\alias{model_predict,pareto_scale,DatasetExperiment-method}
\alias{model_predict,pqn_norm,DatasetExperiment-method}
\alias{model_predict,SVM,DatasetExperiment-method}
\alias{model_predict,vec_norm,DatasetExperiment-method}
\title{Model prediction}
\usage{
\S4method{model_predict}{DFA,DatasetExperiment}(M, D)

\S4method{model_predict}{PCA,DatasetExperiment}(M, D)

\S4method{model_predict}{PLSR,DatasetExperiment}(M, D)

\S4method{model_predict}{PLSDA,DatasetExperiment}(M, D)

\S4method{model_predict}{autoscale,DatasetExperiment}(M, D)

\S4method{model_predict}{blank_filter,DatasetExperiment}(M, D)

\S4method{model_predict}{constant_sum_norm,DatasetExperiment}(M, D)

\S4method{model_predict}{dratio_filter,DatasetExperiment}(M, D)

\S4method{model_predict}{filter_by_name,DatasetExperiment}(M, D)

\S4method{model_predict}{filter_na_count,DatasetExperiment}(M, D)

\S4method{model_predict}{filter_smeta,DatasetExperiment}(M, D)

\S4method{model_predict}{glog_transform,DatasetExperiment}(M, D)

\S4method{model_predict}{linear_model,DatasetExperiment}(M, D)

\S4method{model_predict}{mean_centre,DatasetExperiment}(M, D)

\S4method{model_predict}{mv_feature_filter,DatasetExperiment}(M, D)

\S4method{model_predict}{mv_sample_filter,DatasetExperiment}(M, D)

\S4method{model_predict}{pareto_scale,DatasetExperiment}(M, D)

\S4method{model_predict}{pqn_norm,DatasetExperiment}(M, D)

\S4method{model_predict}{SVM,DatasetExperiment}(M, D)

\S4method{model_predict}{vec_norm,DatasetExperiment}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a DatasetExperiment object}
}
\value{
Returns a modified model object
}
\description{
Apply a model using the input DatasetExperiment. Assumes the model is trained
first.
}
\examples{
M = example_model()
M = model_predict(M,iris_DatasetExperiment())
}
