% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_centre_class.R, R/model_reverse_doc.R
\name{model_reverse,mean_centre,DatasetExperiment-method}
\alias{model_reverse,mean_centre,DatasetExperiment-method}
\alias{model_reverse}
\title{Reverse preprocessing}
\usage{
\S4method{model_reverse}{mean_centre,DatasetExperiment}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a DatasetExperiment object}

\item{...}{additional slots and values passed to struct_class}
}
\value{
Returns a modified DatasetExperiment object
}
\description{
Reverse the effect of a preprocessing step on a DatasetExperiment.
}
\examples{
M = example_model()
D = model_reverse(M,iris_DatasetExperiment())
}
