% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_class.R, R/PCA_class.R, R/PLSR_class.R,
%   R/PLSDA_class.R, R/autoscale_class.R, R/blank_filter_class.R,
%   R/constant_sum_norm_class.R, R/d_ratio_filter_class.R,
%   R/filter_by_name_class.R, R/filter_na_count.R, R/filter_smeta_class.R,
%   R/glog_class.R, R/linear_model_class.R, R/mean_centre_class.R,
%   R/model_train_doc.R, R/mv_feature_filter_class.R,
%   R/mv_sample_filter_class.R, R/paretoscale_class.R,
%   R/pqn_norm_method_class.R, R/svm_classifier_class.R, R/vec_norm_class.R
\name{model_train,DFA,DatasetExperiment-method}
\alias{model_train,DFA,DatasetExperiment-method}
\alias{model_train,PCA,DatasetExperiment-method}
\alias{model_train,PLSR,DatasetExperiment-method}
\alias{model_train,PLSDA,DatasetExperiment-method}
\alias{model_train,autoscale,DatasetExperiment-method}
\alias{model_train,blank_filter,DatasetExperiment-method}
\alias{model_train,constant_sum_norm,DatasetExperiment-method}
\alias{model_train,dratio_filter,DatasetExperiment-method}
\alias{model_train,filter_by_name,DatasetExperiment-method}
\alias{model_train,filter_na_count,DatasetExperiment-method}
\alias{model_train,filter_smeta,DatasetExperiment-method}
\alias{model_train,glog_transform,DatasetExperiment-method}
\alias{model_train,linear_model,DatasetExperiment-method}
\alias{model_train,mean_centre,DatasetExperiment-method}
\alias{model_train}
\alias{model_train,mv_feature_filter,DatasetExperiment-method}
\alias{model_train,mv_sample_filter,DatasetExperiment-method}
\alias{model_train,pareto_scale,DatasetExperiment-method}
\alias{model_train,pqn_norm,DatasetExperiment-method}
\alias{model_train,SVM,DatasetExperiment-method}
\alias{model_train,vec_norm,DatasetExperiment-method}
\title{Train a model}
\usage{
\S4method{model_train}{DFA,DatasetExperiment}(M, D)

\S4method{model_train}{PCA,DatasetExperiment}(M, D)

\S4method{model_train}{PLSR,DatasetExperiment}(M, D)

\S4method{model_train}{PLSDA,DatasetExperiment}(M, D)

\S4method{model_train}{autoscale,DatasetExperiment}(M, D)

\S4method{model_train}{blank_filter,DatasetExperiment}(M, D)

\S4method{model_train}{constant_sum_norm,DatasetExperiment}(M, D)

\S4method{model_train}{dratio_filter,DatasetExperiment}(M, D)

\S4method{model_train}{filter_by_name,DatasetExperiment}(M, D)

\S4method{model_train}{filter_na_count,DatasetExperiment}(M, D)

\S4method{model_train}{filter_smeta,DatasetExperiment}(M, D)

\S4method{model_train}{glog_transform,DatasetExperiment}(M, D)

\S4method{model_train}{linear_model,DatasetExperiment}(M, D)

\S4method{model_train}{mean_centre,DatasetExperiment}(M, D)

\S4method{model_train}{mv_feature_filter,DatasetExperiment}(M, D)

\S4method{model_train}{mv_sample_filter,DatasetExperiment}(M, D)

\S4method{model_train}{pareto_scale,DatasetExperiment}(M, D)

\S4method{model_train}{pqn_norm,DatasetExperiment}(M, D)

\S4method{model_train}{SVM,DatasetExperiment}(M, D)

\S4method{model_train}{vec_norm,DatasetExperiment}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a DatasetExperiment object}
}
\value{
Returns a modified model object
}
\description{
Trains a model using the input DatasetExperiment
}
\examples{
M = example_model()
M = model_train(M,iris_DatasetExperiment())
}
