% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{mv_boxplot-class}
\alias{mv_boxplot-class}
\alias{mv_boxplot}
\title{mv_boxplot class}
\description{
boxplots indicating the numbers of missing values per sample/feature
}
\examples{
C = mv_boxplot()
}
