% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{mv_boxplot}
\alias{mv_boxplot}
\title{mv_boxplot class}
\usage{
mv_boxplot(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
Boxplot of the numbers of missing values per sample/feature
}
\section{Slots}{

\describe{
\item{\code{label_outliers}}{[TRUE] or FALSE to label outliers on the plot
plot}

\item{\code{by_sample}}{by_sample [TRUE] to plot by sample or FALSE to plot by features}

\item{\code{factor_name}}{the sample_meta column to use}

\item{\code{show_counts}}{[TRUE] or FALSE to include the number of samples on the plot}
}}

\examples{
D = sbcms_DatasetExperiment()
C = mv_boxplot(factor_name='class')
chart_plot(C,D)

}
