% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_feature_filter_class.R
\docType{class}
\name{mv_feature_filter-class}
\alias{mv_feature_filter-class}
\alias{mv_feature_filter}
\title{filter features by fraction missing values}
\description{
filters features by the percent number of missing values
}
