% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_feature_filter_class.R
\docType{class}
\name{mv_feature_filter.hist-class}
\alias{mv_feature_filter.hist-class}
\alias{mv_feature_filter.hist}
\title{plot for missing value sample filter}
\description{
plots a histogram of % missing values per sample
}
