% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_feature_filter_class.R
\name{mv_feature_filter_hist}
\alias{mv_feature_filter_hist}
\title{Histogram of missing values per feature}
\usage{
mv_feature_filter_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
mv_feature_filter_hist
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A histogram of the proportion of missing values per feature.
}
\examples{
C = mv_feature_filter_hist()
}
