% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nroot_transform_class.R
\name{nroot_transform}
\alias{nroot_transform}
\title{nth root transform}
\usage{
nroot_transform(root = 2, ...)
}
\arguments{
\item{root}{(numeric) The nth root used for the transform. The default is \code{2}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{nroot_transform} object.
}
\description{
All values in the data matrix are transformed by raising them to the power of 1/n.
}
\examples{
M = nroot_transform()
}
