% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ontology_cache}
\alias{ontology_cache}
\title{ontology cache}
\usage{
ontology_cache()
}
\value{
list of cached ontology terms
}
\description{
A cached list of ontology terms obtained from the ontology lookup service
(OLS) for ontology terms specified for objects in \code{structToolbox}.
}
\examples{
cache = ontology_cache()
}
\seealso{
ontology
}
