% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoscale_class.R
\docType{class}
\name{pareto_scale-class}
\alias{pareto_scale-class}
\alias{pareto_scale}
\title{Pareto scaling}
\value{
A STRUCT model object with methods for pareto scaling.
}
\description{
Pareto scaling centres the columns of the data in a dataset object and divides by the square root of the standard deviation.
}
\examples{
D = iris_dataset()
M = pareto_scale()
M = model.train(M,D)
M = model.predict(M,D)

}
