% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_class.R
\docType{class}
\name{PCA-class}
\alias{PCA-class}
\alias{PCA}
\title{PCA model class}
\description{
Principal Component Analysis (PCA) model class. This object can be used to train/apply PCA mdoels to dataset objects.
}
\examples{
M = PCA()
}
