% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\docType{class}
\name{PCA.dstat-class}
\alias{PCA.dstat-class}
\alias{PCA.dstat}
\title{pca_dstat_plot class}
\description{
line plot showing percent variance and cumulative peercent variance for the computed components.
}
