% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_class.R
\name{PCA}
\alias{PCA}
\title{PCA model class}
\usage{
PCA(number_components = 2, ...)
}
\arguments{
\item{number_components}{The number of principal components to retain}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Principal Component Analysis (PCA) model class. This object can be used to train/apply PCA mdoels to DatasetExperiment objects.
}
\examples{
M = PCA()
}
