% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\docType{class}
\name{pca_biplot_plot-class}
\alias{pca_biplot_plot-class}
\alias{pca_biplot_plot}
\title{pca_biplot_plot class}
\description{
2d scatter plot of princpal component scores overlaid with principal component loadings.
}
