% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\docType{class}
\name{pca_correlation_plot-class}
\alias{pca_correlation_plot-class}
\alias{pca_correlation_plot}
\title{pca_correlation_plot class}
\description{
plots the correlation between features and selected components.
}
\examples{
C = pca_correlation_plot()
}
