% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_correlation_plot}
\alias{pca_correlation_plot}
\title{pca_correlation_plot class}
\usage{
pca_correlation_plot(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
plots the correlation between features and selected components.
}
\examples{
C = pca_correlation_plot()
}
