% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{PCA_dstat}
\alias{PCA_dstat}
\title{pca_dstat_plot class}
\usage{
PCA_dstat(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
line plot showing percent variance and cumulative peercent variance for the computed components.
}
\examples{
C = PCA_dstat()
}
