% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\docType{class}
\name{pca_scores_plot-class}
\alias{pca_scores_plot-class}
\alias{pca_scores_plot}
\title{pca_scores_plot class}
\description{
2d scatter plot of princpal component scores.
}
\examples{
C = pca_scores_plot()
}
