% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_scree_plot}
\alias{pca_scree_plot}
\title{pca_scree_plot_plot class}
\usage{
pca_scree_plot(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Line plot showing percent variance and cumulative percent variance for the computed components.
}
\examples{
C = pca_scree_plot()
}
