% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\docType{class}
\name{permutation_test-class}
\alias{permutation_test-class}
\alias{permutation_test}
\title{permutation test class}
\description{
Applies a permutation test to a model or model.seq()
}
