% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\name{permutation_test.boxplot}
\alias{permutation_test.boxplot}
\title{permutation_test.boxplot class}
\usage{
permutation_test.boxplot(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots the results of a permutation test as a boxplot
}
\examples{
C = permutation_test.boxplot()
}
