% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\docType{class}
\name{permutation_test.hist-class}
\alias{permutation_test.hist-class}
\alias{permutation_test.hist}
\title{permutation_test.hist class}
\description{
plots the results of a permutation test as histograms
}
