% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\name{permutation_test}
\alias{permutation_test}
\title{Permutation test class}
\usage{
permutation_test(number_of_permutations = 50, factor_name, ...)
}
\arguments{
\item{number_of_permutations}{The number of permutations to run}

\item{factor_name}{The same of the sample_meta column to use}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Applies a permutation test to a model or model_seq(). The input metric is calculated for
all permutations, and can be compared to the results from the unpermuted model to
assess model validity.
}
\examples{
I=permutation_test(factor_name='Species')
}
