% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\name{permutation_test.scatter}
\alias{permutation_test.scatter}
\title{permutation_test.scatter class}
\usage{
permutation_test.scatter(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots the results of a permutation test as histograms
}
\examples{
C = permutation_test.scatter()
}
