% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test2_class.R
\docType{class}
\name{permutation_test2-class}
\alias{permutation_test2-class}
\alias{permutation_test2}
\title{permutation test class}
\description{
Applies a permutation test to a model or model.seq()
}
\examples{
I=permutation_test2()

}
