% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test2_class.R
\name{permutation_test2}
\alias{permutation_test2}
\title{permutation test class}
\usage{
permutation_test2(...)
}
\description{
Applies a permutation test to a model or model_seq()
}
\examples{
I=permutation_test2()

}
