% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\docType{class}
\name{permutation_test_hist-class}
\alias{permutation_test_hist-class}
\alias{permutation_test_hist}
\title{permutation_test_hist class}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots the results of a permutation test as histograms
}
\examples{
C = permutation_test_hist()
}
