% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\name{permutation_test_plot}
\alias{permutation_test_plot}
\title{permutation_test_plot class}
\usage{
permutation_test_plot(style = "boxplot", binwidth = 0.05, ...)
}
\arguments{
\item{style}{The plot style. One of 'boxplot', 'violin', 'histogram', 'density' or 'scatter'.}

\item{binwidth}{Binwidth for the "histogram" style. Ignored for all other styles.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the results of a permutation test.
}
\examples{
C = permutation_test_plot(style='boxplot')
}
