% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_class.R
\docType{class}
\name{PLSDA-class}
\alias{PLSDA-class}
\alias{PLSDA}
\title{PLSDA model class}
\description{
Partial least squares (PLS) discriminant analysis (DA) model class. This object can be used to train/apply PLS models.
}
