% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\name{plsda_regcoeff_plot}
\alias{plsda_regcoeff_plot}
\title{plsda_regcoeff_plot class}
\usage{
plsda_regcoeff_plot(level, ...)
}
\arguments{
\item{level}{the group label to plot regression coefficients for}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the regression coefficients of a PLSDA model.
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre()+PLSDA(factor_name='Species')
M = model_apply(M,D)

C = plsda_regcoeff_plot(level='setosa')
chart_plot(C,M[2])
}
