% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\docType{class}
\name{plsda_scores_plot-class}
\alias{plsda_scores_plot-class}
\alias{plsda_scores_plot}
\title{plsda_scores_plot class}
\description{
2d scatter plot of plsda component scores.
}
\examples{
C = plsda_scores_plot()
}
