% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\name{plsda_vip_plot}
\alias{plsda_vip_plot}
\title{plsda_vip_plot class}
\usage{
plsda_vip_plot(threshold = 1, level, ...)
}
\arguments{
\item{threshold}{the VIP threshold to plot. Default = 1}

\item{level}{= the group label to plot VIP scores for}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the vip scores of a PLSDA model.
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre()+PLSDA(factor_name='Species')
M = model_apply(M,D)

C = plsda_vip_plot(level='setosa')
chart_plot(C,M[2])
}
