% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_vip_summary_chart.R
\name{plsda_vip_summary_plot}
\alias{plsda_vip_summary_plot}
\title{PLSDA VIP summary plot}
\usage{
plsda_vip_summary_plot(n_features = 50, ...)
}
\arguments{
\item{n_features}{(numeric, integer) The number of features to include in the summary. The default is \code{50}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
plsda_vip_summary_plot
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A plot of the Variable Importance for Projection (VIP) scores for a PLSDA model for the top selected features.
}
\details{
This object makes use of functionality from the following packages:\itemize{\item{\code{pls}}\item{\code{ggplot2}}\item{\code{reshape2}}\item{\code{cowplot}}}
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre()+PLSDA(factor_name='Species')
M = model_apply(M,D)

C = plsda_vip_summary_plot(n_features=30)
chart_plot(C,M[2])
}
\references{
Mevik B, Wehrens R, Liland K (2020).
\emph{pls: Partial Least Squares and Principal Component Regression}.
R package version 2.7-3, \url{https://CRAN.R-project.org/package=pls}.

Wickham H (2016).
\emph{ggplot2: Elegant Graphics for Data Analysis}.
Springer-Verlag New York.
ISBN 978-3-319-24277-4, \url{https://ggplot2.tidyverse.org}.

Wickham H (2007).
``Reshaping Data with the reshape Package.''
\emph{Journal of Statistical Software}, \bold{21}(12), 1--20.
\url{http://www.jstatsoft.org/v21/i12/}.

Wilke C (2020).
\emph{cowplot: Streamlined Plot Theme and Plot Annotations for 'ggplot2'}.
R package version 1.1.1, \url{https://CRAN.R-project.org/package=cowplot}.
}
