% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\docType{class}
\name{PLSR-class}
\alias{PLSR-class}
\alias{PLSR}
\title{PLSR model class}
\description{
Partial least squares (PLS) Regression model class. This object can be used to train/apply PLS models.
}
\examples{
M = PLSR()
}
