% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\docType{class}
\name{plsr_cook_dist-class}
\alias{plsr_cook_dist-class}
\alias{plsr_cook_dist}
\title{plsr_residual_plot class}
\description{
Plots the residuals for a PLSR model
}
\examples{
C = plsr_cook_dist()
}
