% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_cook_dist}
\alias{plsr_cook_dist}
\title{plsr_residual_plot class}
\usage{
plsr_cook_dist(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the residuals for a PLSR model
}
\examples{
C = plsr_cook_dist()
}
