% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\docType{class}
\name{plsr_prediction_plot-class}
\alias{plsr_prediction_plot-class}
\alias{plsr_prediction_plot}
\title{plsr_prediction_plot class}
\description{
plots the true values against the predicted values for a PLSR model.
}
\examples{
C = plsr_prediction_plot()
}
