% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\docType{class}
\name{plsr_qq_plot-class}
\alias{plsr_qq_plot-class}
\alias{plsr_qq_plot}
\title{plsr_qq_plot class}
\description{
plots the quantiles of PLSR residuals against the qualtiles of a normal
distribution
}
\examples{
C = plsr_qq_plot()
}
