% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\docType{class}
\name{plsr_residual_hist-class}
\alias{plsr_residual_hist-class}
\alias{plsr_residual_hist}
\title{plsr_residual_hist class}
\description{
plots a histogram of the residuals
}
\examples{
C = plsr_residual_hist()
}
