% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqn_norm_method_class.R
\name{pqn_norm_hist}
\alias{pqn_norm_hist}
\title{PQN coefficient histogram}
\usage{
pqn_norm_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
pqn_norm_hist
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A histogram of the PQN coefficients for all features
}
\section{Inheritance}{

A \code{pqn_norm_hist} object inherits the following \code{struct} classes: \cr\cr
\code{pqn_norm_hist()} ⭢ \code{chart()} ⭢ \code{struct_class()}
}

\examples{
M = pqn_norm_hist()

C = pqn_norm_hist()
}
