% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_class.R
\docType{class}
\name{r_squared-class}
\alias{r_squared-class}
\alias{r_squared}
\title{Coefficient of determination class}
\description{
Coefficient of determination (r-squared).
}
\examples{
MET = r_squared()

}
