% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_class.R
\name{r_squared}
\alias{r_squared}
\title{Coefficient of determination class}
\usage{
r_squared(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Coefficient of determination (r-squared).
}
\examples{
MET = r_squared()

}
