% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_class.R
\name{resample_chart}
\alias{resample_chart}
\title{resample_chart class}
\usage{
resample_chart(style = "boxplot", binwidth = 0.05, ...)
}
\arguments{
\item{style}{The plot style. One of 'boxplot', 'violin', 'histogram', 'density' or 'scatter'.}

\item{binwidth}{Binwidth for the "histogram" style. Ignored for all other styles.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the results of a resampling.
}
\examples{
C = resample_chart(style='boxplot')
}
