% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_filter.R
\docType{class}
\name{rsd_filter-class}
\alias{rsd_filter-class}
\alias{rsd_filter}
\title{rsd filter}
\description{
filters features based on the relative standard deviation (RSD) for the QC samples
}
\examples{
M = rsd_filter()

}
