% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_filter.R
\docType{class}
\name{rsd_filter.hist-class}
\alias{rsd_filter.hist-class}
\alias{rsd_filter.hist}
\title{plot for rsd filter}
\description{
plots a histogram of the calculated RSD for the RSD filter
}
\examples{
C = rsd_filter.hist()

}
