% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_filter.R
\name{rsd_filter_hist}
\alias{rsd_filter_hist}
\title{RSD histogram}
\usage{
rsd_filter_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
rsd_filter_hist
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A histogram of the calculated RSD values.
}
\examples{
C = rsd_filter_hist()

}
