% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbcms_dataset_class.R
\name{sbcms_DatasetExperiment}
\alias{sbcms_DatasetExperiment}
\title{Data from sbcms}
\usage{
sbcms_DatasetExperiment(filtered = FALSE)
}
\arguments{
\item{filtered}{TRUE to load prefiltered data, or FALSE to load the unfiltered data}

\item{...}{slots and values for the new object}
}
\value{
struct object

DatasetExperiment object
}
\description{
SBCMS package data after correction, as a struct DatasetExperiment object
}
\examples{
D = sbcms_DatasetExperiment()
summary(D)
}
