% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_class.R
\name{split_data}
\alias{split_data}
\title{Split data into subsets}
\usage{
split_data(p, ...)
}
\arguments{
\item{p}{The proportion of samples in the training set.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Splits the data into a training and test set.
}
\examples{
M = split_data(p=0.75)

}
