% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_split_class.R
\name{stratified_split}
\alias{stratified_split}
\title{Stratified sampling}
\usage{
stratified_split(p_train, factor_name, ...)
}
\arguments{
\item{p_train}{The proportion of samples in the training set.}

\item{factor_name}{The column of sample_meta to use for stratification}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Splits the data into a training and test set, using stratification to keep group sizes
in equal proportions to the full dataset.
}
\examples{
D = iris_DatasetExperiment()
M = stratified_split(p_train=0.75,factor_name='Species')
M = model_apply(M,D)

}
