% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structtoolbox.R
\docType{package}
\name{struct.toolbox}
\alias{struct.toolbox}
\alias{struct.toolbox-package}
\title{struct.toolbox: Examples of tools built using the Statistics in R Using Class Templates (struct) package}
\description{
This package extends the classes defined in the struct package
}
