% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structToolbox.R
\docType{package}
\name{structToolbox-package}
\alias{structToolbox}
\alias{structToolbox-package}
\title{structToolbox: Examples of tools built using the Statistics in R Using Class Templates (struct) package}
\description{
This package extends the classes defined in the struct package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/computational-metabolomics/structToolbox}
  \item \url{https://computational-metabolomics.github.io/structToolbox/}
}

}
\author{
\strong{Maintainer}: Gavin Rhys Lloyd \email{g.r.lloyd@bham.ac.uk} (\href{https://orcid.org/0000-0001-7989-6695}{ORCID})

Authors:
\itemize{
  \item Ralf Johannes Maria Weber \email{r.j.weber@bham.ac.uk}
}

}
\keyword{internal}
