% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_class.R, R/PLSDA_class.R, R/PLSR_class.R,
%   R/autoscale_class.R, R/linear_model_class.R, R/mean_centre_class.R
\docType{methods}
\name{model.train,PCA,dataset-method}
\alias{model.train,PCA,dataset-method}
\alias{model.train,PLSDA,dataset-method}
\alias{model.train,PLSR,dataset-method}
\alias{model.train,autoscale,dataset-method}
\alias{model.train,linear_model,dataset-method}
\alias{model.train,mean_centre,dataset-method}
\title{Train a model}
\usage{
\S4method{model.train}{PCA,dataset}(M, D)

\S4method{model.train}{PLSDA,dataset}(M, D)

\S4method{model.train}{PLSR,dataset}(M, D)

\S4method{model.train}{autoscale,dataset}(M, D)

\S4method{model.train}{linear_model,dataset}(M, D)

\S4method{model.train}{mean_centre,dataset}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}

\item{M}{a model object}

\item{D}{a dataset object}

\item{M}{a model object}

\item{D}{a dataset object}

\item{M}{a model object}

\item{D}{a dataset object}

\item{M}{a model object}

\item{D}{a dataset object}

\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified model object

Returns a modified model object

Returns a modified model object

Returns a modified model object

Returns a modified model object

Returns a modified model object
}
\description{
Trains a model using the input dataset

Trains a model using the input dataset

Trains a model using the input dataset

Trains a model using the input dataset

Trains a model using the input dataset

Trains a model using the input dataset
}
\examples{
M = example_model()
M = model.train(M,iris_dataset())

M = example_model()
M = model.train(M,iris_dataset())

M = example_model()
M = model.train(M,iris_dataset())

M = example_model()
M = model.train(M,iris_dataset())

M = example_model()
M = model.train(M,iris_dataset())

M = example_model()
M = model.train(M,iris_dataset())

}
