% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE_class.R
\docType{class}
\name{tSNE-class}
\alias{tSNE-class}
\alias{tSNE}
\title{tSNE model class}
\description{
t-Distributed Stochastic Neighbor Embedding (tSNE) class. This object can be used to train/apply tSNE models to dataset objects.
}
