% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE_class.R
\name{tSNE}
\alias{tSNE}
\title{tSNE method class}
\usage{
tSNE(
  dims = 2,
  perplexity = 30,
  max_iter = 100,
  theta = 0.5,
  check_duplicates = FALSE,
  init = NULL,
  eta = 200,
  ...
)
}
\arguments{
\item{dims}{integer; Output dimensionality (default: 2)}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{max_iter}{integer; Number of iterations (default: 1000)}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.5)}

\item{check_duplicates}{logical; Checks whether duplicates are present. It is best to make sure there are no duplicates present and set this option to FALSE, especially for large datasets (default: TRUE)}

\item{init}{Initial locations of the objects. If NULL, random initialization will be used.
If NULL then}

\item{eta}{numeric; Learning rate (default: 200.0)}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
t-Distributed Stochastic Neighbor Embedding (tSNE) class. This object can be
used to train/apply tSNE models to DatasetExperiment objects.
}
\details{
This object is a wrapper for Rtsne::Rtsne.
}
\examples{
M = tSNE()

}
