% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_norm_class.R
\name{vec_norm}
\alias{vec_norm}
\title{vector nromalisation}
\usage{
vec_norm(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
applies vector normalisation
}
\examples{
M = vec_norm()

}
