% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test_class.R
\docType{class}
\name{wilcox_test-class}
\alias{wilcox_test-class}
\alias{wilcox_test}
\title{wilcoxon signed rank test method class}
\description{
wilcoxon signed rank test. Calculate signed rank test for all features in a
dataset. Non-parametric ttest.
}
