% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_class.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA}
\usage{
ANOVA(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
A struct method object with functions for applying ANOVA
}
\description{
Applies ANOVA to each feature in a DatasetExperiment object.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{p-value threshold for determining significance. Default alpha = 0.05.}

\item{\code{mtc}}{multiple test correction method to apply. Can be: holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}

\item{\code{formula}}{the formula to use in the form y~...}
}}

\examples{
D = iris_DatasetExperiment()
M = ANOVA(formula=y~Species)
M = model_apply(M,D)

}
