% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_class.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA}
\usage{
ANOVA(alpha = 0.05, mtc = "fdr", formula, ss_type = "III")
}
\arguments{
\item{alpha}{The p-value threshold. Default alpha = 0.05.}

\item{mtc}{Multiple test correction method passed to \code{p.adjust}. Default mtc = 'fdr'.}

\item{formula}{The formula to use for ANOVA. See \code{lm} for details.}

\item{ss_type}{The type of sum of squares to use. Can be I, II or III. Default is ss_type = 'III'.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
ANOVA object
}
\description{
Applies ANOVA to each feature in a DatasetExperiment object.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{p-value threshold for determining significance. Default alpha = 0.05.}

\item{\code{mtc}}{multiple test correction method to apply. Can be: holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}

\item{\code{formula}}{the formula to use in the form y~...}
}}

\examples{
D = iris_DatasetExperiment()
M = ANOVA(formula=y~Species)
M = model_apply(M,D)

}
