% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscale_class.R
\docType{class}
\name{autoscale-class}
\alias{autoscale-class}
\alias{autoscale}
\title{Autoscale}
\value{
A STRUCT model object with methods for autoscaling.
}
\description{
Autoscaling centres the columns of the data in a dataset object and divides by the standard deviation.
}
\examples{
D = iris_dataset()
M = autoscale()
M = model.train(M,D)
M = model.predict(M,D)

}
