% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_accuracy_class.R
\name{balanced_accuracy}
\alias{balanced_accuracy}
\title{Balanced Accuracy}
\usage{
balanced_accuracy(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{balanced_accuracy} object.
}
\description{
Balanced Accuracy is the average proportion of correctly classified samples across all groups.
}
\examples{
D = iris_DatasetExperiment()
XCV = kfold_xval(folds=5,factor_name='Species') *
      (mean_centre() + PLSDA(number_components=2,factor_name='Species'))
MET = balanced_accuracy()
XCV = run(XCV,D,MET)
}
