% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_filter_class.R
\name{blank_filter}
\alias{blank_filter}
\title{Blank filter}
\usage{
blank_filter(
  fold_change = 20,
  blank_label = "blank",
  qc_label = "QC",
  factor_name,
  fraction_in_blank = 0,
  ...
)
}
\arguments{
\item{fold_change}{\code{numeric(1)}, fold_change minimum fold change 
between analytical and blank samples.}

\item{blank_label}{\code{character(1)}, class label used to identify blank
samples.}

\item{qc_label}{\code{character(1)} or \code{NULL}, class label used to
identify QC samples.}

\item{factor_name}{the column name of sample_meta to use}

\item{fraction_in_blank}{\code{numeric(1)}, value between 0 to 1 to specify 
fraction in how many blanks peaks should be present.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
A struct model object for applying a blank filter
}
\description{
Filters features based on the features present in blank samples. The median
intensity of the samples is compared to the median intensity of the blank
samples. Any sample not sufficiently more intense than the blank is removed.
This is a wrapper for the blank filter in the PMP package.
}
\examples{
D = iris_DatasetExperiment()
M = blank_filter(fold_change=2,
                 factor_name='Species',
                 blank_label='setosa',
                 qc_label='versicolor')
M = model_apply(M,D)
}
