% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_filter_class.R
\name{blank_filter_hist}
\alias{blank_filter_hist}
\title{plot for blank filter}
\usage{
blank_filter_hist(...)
}
\description{
plots a histogram of the calculated fold change for the blank filter (median blank / median sample)
}
\examples{
C = blank_filter_hist()
}
