% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_filter_class.R
\name{blank_filter_hist}
\alias{blank_filter_hist}
\title{Histogram of blank filter fold changes}
\usage{
blank_filter_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
blank_filter_hist
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A histogram of the calculated fold changes for the blank filter (median samples divided by median blanks)
}
\examples{
C = blank_filter_hist()
}
