% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_class.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap class}
\usage{
bootstrap(number_of_iterations = 100, collect, ...)
}
\arguments{
\item{number_of_iterations}{The number of bootstrap iterations to run}

\item{collect}{The name of model output to collect over all iterations}

\item{...}{additional slots and values passed to struct_class}
}
\value{
A struct iterator object
}
\description{
Applies bootstrapping to a model or model_seq(). Any output from the model can
be 'collected' over all bootstrap repetitions for further analysis.
}
\examples{
D = iris_DatasetExperiment()
I = bootstrap(number_of_iterations = 5, collect='vip') *
    (mean_centre() + PLSDA(factor_name = 'Species'))
I = run(I,D,balanced_accuracy())

}
