% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_class.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap resampling}
\usage{
bootstrap(number_of_repetitions = 100, collect, ...)
}
\arguments{
\item{number_of_repetitions}{(numeric, integer) The number of bootstrap repetitions. The default is \code{100}.}

\item{collect}{(character) The name of a model output to collect over all bootstrap repetitions, in addition to the input metric.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{bootstrap} object.
}
\description{
In bootstrap resampling a subset of samples is selected at random with replacement to form a training set. Any sample not selected for training is included in the test set. This process is repeated many times, and performance metrics are computed for each repetition.
}
