% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_accuracy_class.R, R/calculate_doc.R,
%   R/r_squared_class.R
\docType{methods}
\name{calculate,balanced_accuracy-method}
\alias{calculate,balanced_accuracy-method}
\alias{calculate}
\alias{calculate,r_squared-method}
\title{Calculate metric}
\usage{
\S4method{calculate}{balanced_accuracy}(obj, Y, Yhat)

\S4method{calculate}{r_squared}(obj, Y, Yhat)
}
\arguments{
\item{obj}{a metric object}

\item{value}{value}

\item{...}{additional inputs depending on object}
}
\description{
Calculate metric
}
\examples{
MET = metric()
calculate(MET)
}
