% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_class.R, R/PCA_plotfcns.R,
%   R/PLSDA_charts.R, R/PLSR_class.R, R/blank_filter_class.R,
%   R/chart_plot_doc.R, R/confounders_clsq_class.R,
%   R/dataset_chart_classes.R, R/factor_barchart_class.R,
%   R/feature_plot_array_class.R, R/feature_profile_class.R,
%   R/fold_change_class.R, R/forward_selection_by_rank_class.R,
%   R/glog_class.R, R/grid_search_1d_class.R, R/hca_class.R,
%   R/kfold_xval_charts.R, R/kw_rank_sum_class.R, R/mv_feature_filter_class.R,
%   R/mv_sample_filter_class.R, R/permutation_test_class.R,
%   R/pqn_norm_method_class.R, R/rsd_filter.R, R/svm_classifier_class.R,
%   R/tSNE_class.R, R/wilcox_test_class.R
\name{chart_plot,dfa_scores_plot,DFA-method}
\alias{chart_plot,dfa_scores_plot,DFA-method}
\alias{chart_plot,pca_correlation_plot,PCA-method}
\alias{chart_plot,pca_scores_plot,PCA-method}
\alias{chart_plot,pca_biplot,PCA-method}
\alias{chart_plot,pca_loadings_plot,PCA-method}
\alias{chart_plot,pca_scree_plot,PCA-method}
\alias{chart_plot,pca_dstat_plot,PCA-method}
\alias{chart_plot,plsda_scores_plot,PLSDA-method}
\alias{chart_plot,plsda_predicted_plot,PLSDA-method}
\alias{chart_plot,plsda_roc_plot,PLSDA-method}
\alias{chart_plot,plsda_vip_plot,PLSDA-method}
\alias{chart_plot,plsda_regcoeff_plot,PLSDA-method}
\alias{chart_plot,plsr_prediction_plot,PLSR-method}
\alias{chart_plot,plsr_residual_hist,PLSR-method}
\alias{chart_plot,plsr_qq_plot,PLSR-method}
\alias{chart_plot,plsr_cook_dist,PLSR-method}
\alias{chart_plot,blank_filter_hist,blank_filter-method}
\alias{chart_plot}
\alias{chart_plot,confounders_lsq_barchart,confounders_clsq-method}
\alias{chart_plot,confounders_lsq_boxplot,confounders_clsq-method}
\alias{chart_plot,feature_boxplot,DatasetExperiment-method}
\alias{chart_plot,mv_histogram,DatasetExperiment-method}
\alias{chart_plot,mv_boxplot,DatasetExperiment-method}
\alias{chart_plot,DatasetExperiment_dist,DatasetExperiment-method}
\alias{chart_plot,DatasetExperiment_boxplot,DatasetExperiment-method}
\alias{chart_plot,compare_dist,DatasetExperiment-method}
\alias{chart_plot,DatasetExperiment_heatmap,DatasetExperiment-method}
\alias{chart_plot,DatasetExperiment_factor_boxplot,DatasetExperiment-method}
\alias{chart_plot,feature_profile,DatasetExperiment-method}
\alias{chart_plot,fold_change_plot,fold_change-method}
\alias{chart_plot,fs_line,forward_selection_by_rank-method}
\alias{chart_plot,glog_opt_plot,glog_transform-method}
\alias{chart_plot,gs_line,grid_search_1d-method}
\alias{chart_plot,hca_dendrogram,HCA-method}
\alias{chart_plot,kfoldxcv_grid,kfold_xval-method}
\alias{chart_plot,kfoldxcv_metric,kfold_xval-method}
\alias{chart_plot,kw_p_hist,kw_rank_sum-method}
\alias{chart_plot,mv_feature_filter_hist,mv_feature_filter-method}
\alias{chart_plot,mv_sample_filter_hist,mv_sample_filter-method}
\alias{chart_plot,permutation_test_plot,permutation_test-method}
\alias{chart_plot,pqn_norm_hist,pqn_norm-method}
\alias{chart_plot,rsd_filter_hist,rsd_filter-method}
\alias{chart_plot,svm_plot_2d,SVM-method}
\alias{chart_plot,tSNE_scatter,tSNE-method}
\alias{chart_plot,wilcox_p_hist,wilcox_test-method}
\title{chart_plot method}
\usage{
\S4method{chart_plot}{dfa_scores_plot,DFA}(obj, dobj)

\S4method{chart_plot}{pca_correlation_plot,PCA}(obj, dobj)

\S4method{chart_plot}{pca_scores_plot,PCA}(obj, dobj)

\S4method{chart_plot}{pca_biplot,PCA}(obj, dobj)

\S4method{chart_plot}{pca_loadings_plot,PCA}(obj, dobj)

\S4method{chart_plot}{pca_scree_plot,PCA}(obj, dobj)

\S4method{chart_plot}{pca_dstat_plot,PCA}(obj, dobj)

\S4method{chart_plot}{plsda_scores_plot,PLSDA}(obj, dobj)

\S4method{chart_plot}{plsda_predicted_plot,PLSDA}(obj, dobj)

\S4method{chart_plot}{plsda_roc_plot,PLSDA}(obj, dobj)

\S4method{chart_plot}{plsda_vip_plot,PLSDA}(obj, dobj)

\S4method{chart_plot}{plsda_regcoeff_plot,PLSDA}(obj, dobj)

\S4method{chart_plot}{plsr_prediction_plot,PLSR}(obj, dobj)

\S4method{chart_plot}{plsr_residual_hist,PLSR}(obj, dobj)

\S4method{chart_plot}{plsr_qq_plot,PLSR}(obj, dobj)

\S4method{chart_plot}{plsr_cook_dist,PLSR}(obj, dobj)

\S4method{chart_plot}{blank_filter_hist,blank_filter}(obj, dobj)

\S4method{chart_plot}{confounders_lsq_barchart,confounders_clsq}(obj, dobj)

\S4method{chart_plot}{confounders_lsq_boxplot,confounders_clsq}(obj, dobj)

\S4method{chart_plot}{feature_boxplot,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{mv_histogram,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{mv_boxplot,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{DatasetExperiment_dist,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{DatasetExperiment_boxplot,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{compare_dist,DatasetExperiment}(obj, dobj, eobj)

\S4method{chart_plot}{DatasetExperiment_heatmap,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{DatasetExperiment_factor_boxplot,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{feature_profile,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{fold_change_plot,fold_change}(obj, dobj)

\S4method{chart_plot}{fs_line,forward_selection_by_rank}(obj, dobj)

\S4method{chart_plot}{glog_opt_plot,glog_transform}(obj, dobj, gobj)

\S4method{chart_plot}{gs_line,grid_search_1d}(obj, dobj)

\S4method{chart_plot}{hca_dendrogram,HCA}(obj, dobj)

\S4method{chart_plot}{kfoldxcv_grid,kfold_xval}(obj, dobj)

\S4method{chart_plot}{kfoldxcv_metric,kfold_xval}(obj, dobj)

\S4method{chart_plot}{kw_p_hist,kw_rank_sum}(obj, dobj)

\S4method{chart_plot}{mv_feature_filter_hist,mv_feature_filter}(obj, dobj)

\S4method{chart_plot}{mv_sample_filter_hist,mv_sample_filter}(obj, dobj)

\S4method{chart_plot}{permutation_test_plot,permutation_test}(obj, dobj)

\S4method{chart_plot}{pqn_norm_hist,pqn_norm}(obj, dobj)

\S4method{chart_plot}{rsd_filter_hist,rsd_filter}(obj, dobj)

\S4method{chart_plot}{svm_plot_2d,SVM}(obj, dobj, gobj)

\S4method{chart_plot}{tSNE_scatter,tSNE}(obj, dobj)

\S4method{chart_plot}{wilcox_p_hist,wilcox_test}(obj, dobj)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a struct object}

\item{eobj}{a second DatasetExperiment object to compare with the first}

\item{gobj}{The DatasetExperiment object used with glog_transform}
}
\value{
a plot object
}
\description{
Plots a chart object
}
\examples{
C = example_chart()
chart_plot(C,iris_DatasetExperiment())
}
