% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_lsq_class.R
\docType{class}
\name{classical_lsq-class}
\alias{classical_lsq-class}
\alias{classical_lsq}
\title{univariate classical least squares regression}
\description{
classical least squares, where y is the response and x is the design matrix, applied to each feature individually.
}
\examples{
M = classical_lsq()
}
