% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_lsq_class.R
\name{classical_lsq}
\alias{classical_lsq}
\title{Classical Least Squares regression}
\usage{
classical_lsq(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
A STRUCT method object with functions for applying classical least squares
}
\description{
Classical least squares, where y is the response and X is the design matrix,
applied to each feature individually. Here the response is taken from the
data matrix and the design matrix is the taken from the specified sample meta
data column.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{p-value threshold for determining significance. Default alpha = 0.05.}

\item{\code{mtc}}{multiple test correction method to apply. Can be: holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}

\item{\code{factor_names}}{the column name(s) of sample_meta to use}

\item{\code{intercept}}{[TRUE] or FALSE to include an intercept term in the fit}
}}

\examples{
D = iris_DatasetExperiment()
M = classical_lsq(factor_names = 'Species')
M = model_apply(M,D)

}
