% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{compare_dist}
\alias{compare_dist}
\title{Compare distributions}
\usage{
compare_dist(factor_name, ...)
}
\arguments{
\item{factor_name}{the sample_meta colum to use}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
A combination of plots to compare distributions of samples/features in two
datasets
}
\examples{
D1=MTBLS79_DatasetExperiment(filtered=FALSE)
D2=MTBLS79_DatasetExperiment(filtered=TRUE)
C = compare_dist(factor_name='class')
chart_plot(C,D1,D2)
}
