% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{compare_dist}
\alias{compare_dist}
\title{Compare distributions}
\usage{
compare_dist(factor_name, ...)
}
\arguments{
\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{compare_dist} object.
}
\description{
Histograms and boxplots computed across samples and features are used to visually compare two datasets e.g. before and after filtering and/or normalisation.
}
\examples{
D1=MTBLS79_DatasetExperiment(filtered=FALSE)
D2=MTBLS79_DatasetExperiment(filtered=TRUE)
C = compare_dist(factor_name='class')
chart_plot(C,D1,D2)
}
