% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_coef_class.R
\name{corr_coef}
\alias{corr_coef}
\title{correlation model class}
\usage{
corr_coef(...)
}
\description{
correlation model class. Calculate correlation between features and continuous variables
}
\examples{
M = corr_coef()

}
