% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{dataset.boxplot-class}
\alias{dataset.boxplot-class}
\alias{dataset.boxplot}
\title{distribution plot class}
\description{
visualise distributions of features/samples
}
\examples{
C = dataset.boxplot()
}
