% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_barchart_class.R
\docType{class}
\name{dataset.factor_barchart-class}
\alias{dataset.factor_barchart-class}
\alias{dataset.factor_barchart}
\title{dataset.factor_barchart chart class}
\description{
Bar charts based on groupings by factor.
}
\examples{
D = iris_dataset()
C = dataset.factor_barchart(factor_names='Species',feature_to_plot='Petal.Width')
chart.plot(C,D)
}
