% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_barchart_class.R
\name{DatasetExperiment_factor_boxplot}
\alias{DatasetExperiment_factor_boxplot}
\title{Factor boxplot}
\usage{
DatasetExperiment_factor_boxplot(feature_to_plot, factor_names, ...)
}
\arguments{
\item{feature_to_plot}{(character, numeric, integer) The name of the plotted feature.}

\item{factor_names}{(character) The name of sample meta column(s) to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
DatasetExperiment_factor_boxplot
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
Boxplot for a feature to visualise the distribution of values within each group
}
\section{Inheritance}{

A \code{DatasetExperiment_factor_boxplot} object inherits the following \code{struct} classes: \cr\cr
\code{DatasetExperiment_factor_boxplot()} ⭢ \code{chart()} ⭢ \code{struct_class()}
}

\examples{
M = DatasetExperiment_factor_boxplot(
      factor_names = "V1",
      feature_to_plot = "V1")

D = iris_DatasetExperiment()
C = DatasetExperiment_factor_boxplot(factor_names='Species',feature_to_plot='Petal.Width')
chart_plot(C,D)
}
