% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{DatasetExperiment_heatmap}
\alias{DatasetExperiment_heatmap}
\title{DatasetExperiment_heatmap class}
\usage{
DatasetExperiment_heatmap(na_colour = "#FF00E4", ...)
}
\arguments{
\item{na_colour}{A hex colour code to use for missing values}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots a DatasetExperiment as a heatmap
}
\examples{
C = DatasetExperiment_heatmap()
}
