% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_class.R
\name{DFA}
\alias{DFA}
\title{Discriminant Factor Analysis (DFA)}
\usage{
DFA(factor_name, ...)
}
\arguments{
\item{factor_name}{The sample_meta column name containing group labels}

\item{...}{additional slots and values passed to struct_class}

\item{number_factors}{The number of discriminant factors to calculate}
}
\value{
struct object
}
\description{
Applies Dicriminant Factor Analysis to a dataset.
}
\examples{
D = iris_DatasetExperiment()
M = DFA(factor_name='Species')
M = model_apply(M,D)
}
