% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_class.R
\name{dfa_scores_plot}
\alias{dfa_scores_plot}
\title{dfa_scores_plot class}
\usage{
dfa_scores_plot(
  components = c(1, 2),
  points_to_label = "none",
  factor_name,
  ellipse = "all",
  label_filter = character(0),
  label_factor = "rownames",
  label_size = 3.88,
  ...
)
}
\arguments{
\item{components}{The discriminant factors to plot (\code{numeric(2)})}

\item{points_to_label}{"none", "all", or "outliers" will be labelled on the plot.}

\item{factor_name}{The sample_meta column name to use for colouring the points.
You can provide up to two factors for this plot.}

\item{ellipse}{"all" will plot all ellipses, "group" will only plot group ellipses,
"none" will not plot any ellipses and "sample" will plot ellipse for all samples (ignoring group).}

\item{label_filter}{Only include labels for samples in the group specified by label_filter.
If zero length then all labels will be included.}

\item{label_factor}{The sample_meta column to use for labelling the samples.
If 'rownames' then the rownames will be used.}

\item{label_size}{The text size of the labels.NB ggplot units, not font size units.
Default 3.88.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
2d scatter plot of discriminant factor scores.
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre() + DFA(factor_name='Species')
M = model_apply(M,D)
C = dfa_scores_plot(factor_name = 'Species')
chart_plot(C,M[2])

}
