% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ratio_filter_class.R
\name{dratio_filter}
\alias{dratio_filter}
\title{Dispersion ratio filter}
\usage{
dratio_filter(threshold = 20, qc_label = "QC", factor_name, ...)
}
\arguments{
\item{threshold}{(numeric) The threshold below which features are removed. The default is \code{20}.}

\item{qc_label}{(character) The label used to identify QC samples. The default is \code{"QC"}.}

\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{dratio_filter} object.
}
\description{
The dispersion ratio (d-ratio) compares the standard deviation (or non-parametric equivalent) of the Quality Control (QC) samples relative to the standard deviation (or non-parametric equivalent) of the samples for each feature. If the d-ratio is greater than a predefined threshold then the observed sample variance could be due to technical variance and the feature is removed.
}
\examples{
D = MTBLS79_DatasetExperiment()
M = dratio_filter(threshold=20,qc_label='QC',factor_name='Class')
M = model_apply(M,D)
}
\references{
Broadhurst D, Goodacre R, Reinke SN, Kuligowski J, Wilson ID, Lewis MR, Dunn WB (2018).
``Guidelines and considerations for the use of system suitability and quality control samples in mass spectrometry assays applied in untargeted clinical metabolomic studies.''
\emph{Metabolomics}, \bold{14}(6).
}
