% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ratio_filter_class.R
\name{dratio_filter}
\alias{dratio_filter}
\title{D ratio filter}
\usage{
dratio_filter(...)
}
\arguments{
\item{threshold}{D ratio threshold. Features with a d-ratio larger than this
value are removed.}

\item{qc_label}{the label used to identify QC labels}

\item{factor_name}{the the sample_meta data column containing the QC labels}
}
\value{
A struct method object with functions for filtering using the d-ratio.
}
\description{
Filters features based on their D ratio, which is the ratio of technical to
sample variance.
}
\examples{
D = sbcms_DatasetExperiment()
M = dratio_filter(threshold=20,qc_label='QC',factor_name='class')
M = model_apply(M,D)

}
