% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{feature_boxplot-class}
\alias{feature_boxplot-class}
\alias{feature_boxplot}
\title{feature_boxplot class}
\description{
plots a boxplot of a chosen feature for each group of a dataset
}
