% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{feature_boxplot}
\alias{feature_boxplot}
\title{Feature boxplots}
\usage{
feature_boxplot(
  label_outliers = TRUE,
  feature_to_plot,
  factor_name,
  show_counts = TRUE,
  ...
)
}
\arguments{
\item{label_outliers}{[TRUE] or FALSE to label outliers on the plot}

\item{feature_to_plot}{the column id to plot.}

\item{factor_name}{the sample_meta column to use}

\item{show_counts}{[TRUE] or FALSE to include the number of samples on the
plot}

\item{...}{additional slots and values passed to struct_class}
}
\value{
A struct chart object
}
\description{
Plots a boxplot of a chosen feature for each group of a input factor.
}
\examples{
D = MTBLS79_DatasetExperiment
C = feature_boxplot(factor_name='Species',feature_to_plot='Petal.Width')
chart_plot(C,D)
}
