% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{feature_boxplot}
\alias{feature_boxplot}
\title{feature_boxplot class}
\usage{
feature_boxplot(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
plots a boxplot of a chosen feature for each group of a DatasetExperiment
}
\section{Slots}{

\describe{
\item{\code{label_outliers}}{[TRUE] or FALSE to label outliers on the plot}

\item{\code{feature_to_plot}}{the column id to plot.}

\item{\code{factor_name}}{the sample_meta column to use}

\item{\code{show_counts}}{[TRUE] or FALSE to include the number of samples on the
plot}
}}

\examples{
D = sbcms_DatasetExperiment
C = feature_boxplot(factor_name='Species',feature_to_plot='Petal.Width')
chart_plot(C,D)

}
