% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_profile_class.R
\name{feature_profile}
\alias{feature_profile}
\title{Feature profile}
\usage{
feature_profile(
  run_order,
  qc_label,
  qc_column,
  colour_by,
  feature_to_plot,
  plot_sd = FALSE,
  ...
)
}
\arguments{
\item{run_order}{(character) The sample-meta column name containing run order.}

\item{qc_label}{(character) The label used to identify QC samples.}

\item{qc_column}{(character) The sample-meta column name containing the labels used to identify QC samples.}

\item{colour_by}{(character) The sample-meta column name to used to colour the plot.}

\item{feature_to_plot}{(numeric, character, integer) The name or column id of the plotted feature.}

\item{plot_sd}{(logical) Plot standard deviation. Allowed values are limited to the following: \itemize{\item{\code{"TRUE"}: Standard deviation of samples and QCs are included on the plot.}\item{\code{"FALSE"}: Standard deviation is not plotted.}} The default is \code{FALSE}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{feature_profile} object.
}
\description{
A plot visualising the change in intensity of a feature with a continuous variable such as time, dose, or run order.
}
\examples{
D = MTBLS79_DatasetExperiment()
C = feature_profile(run_order='sample_order',
    qc_label='QC',
    qc_column='class',
    colour_by='class',
    feature_to_plot=1)
chart_plot(C,D)
}
