% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_name_class.R
\name{filter_by_name}
\alias{filter_by_name}
\title{Filter by name}
\usage{
filter_by_name(mode = "exclude", dimension = "sample", names, ...)
}
\arguments{
\item{mode}{"include" or ["exclude"] to subsample a DatasetExperiment by including or
excluding samples/features based on the provided labels}

\item{dimension}{["sample"] or "variable" to filter by sample or feature
labels}

\item{names}{the sample/feature identifiers to filter by. Can provide column
names, column indices or logical.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
A filter to subsample a DatasetExperiment object based on sample or feature name,
id, row/column index or using a vector of TRUE/FALSE.
}
\examples{
D = MTBLS79_DatasetExperiment()
M = filter_by_name(mode='exclude',dimension='variable',names=c(1,2,3))
M = model_apply(M,D)

}
