% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_na_count.R
\name{filter_na_count}
\alias{filter_na_count}
\title{filter_na_count class}
\usage{
filter_na_count(...)
}
\arguments{
\item{threshold}{the maximum number of NA allowed per level of factor_name}

\item{factor_name}{the sample_meta column name to use}
}
\description{
Filters features by the number of NA per class
}
\examples{
D = sbcms_DatasetExperiment()
M = filter_na_count(threshold=3,factor_name='class')
M = model_apply(M,D)

}
