% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_smeta_class.R
\docType{class}
\name{filter_smeta-class}
\alias{filter_smeta-class}
\alias{filter_smeta}
\title{filter_smeta class}
\arguments{
\item{mode}{= ['include'] or 'exclude' to include or exclude samples based on
the provided labels}

\item{levels}{a list of level names to include/exclude}

\item{factor_name}{the sample_meta column name to use}
}
\description{
A filter to subset a dataset object based on sample meta data.
}
\examples{
D = sbcms_dataset()
M = filter_smeta(mode='exclude',levels='QC',factor_name='QC')
M = model.apply(M,D)

}
