% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_smeta_class.R
\name{filter_smeta}
\alias{filter_smeta}
\title{filter_smeta class}
\usage{
filter_smeta(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
A filter to subset a DatasetExperiment object based on sample meta data.
}
\section{Slots}{

\describe{
\item{\code{mode}}{= ['include'] or 'exclude' to include or exclude samples based on
the provided labels}

\item{\code{levels}}{a list of level names to include/exclude}

\item{\code{factor_name}}{the sample_meta column name to use}
}}

\examples{
D = sbcms_DatasetExperiment()
M = filter_smeta(mode='exclude',levels='QC',factor_name='QC')
M = model_apply(M,D)

}
