% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_exact_class.R
\docType{class}
\name{fisher_exact-class}
\alias{fisher_exact-class}
\alias{fisher_exact}
\title{Fisher's exact test model class}
\description{
Fisher's exact test class. Applies FET for all features in a dataset
}
\examples{
M = fisher_exact()
}
