% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_class.R
\name{fold_change}
\alias{fold_change}
\title{fold change class}
\usage{
fold_change(
  alpha = 0.05,
  factor_name,
  paired = FALSE,
  sample_name = character(0),
  threshold = 2,
  control_group = character(0),
  method = "geometric",
  ...
)
}
\arguments{
\item{alpha}{confidence level to use for intervals}

\item{factor_name}{the sample_meta column to use}

\item{paired}{TRUE or [FALSE] to account for paired samples}

\item{sample_name}{the sample_meta column name to use for a paired samples}

\item{threshold}{a threshold to define fold change as 'significant'.}

\item{control_group}{a level of factor name to use as the control group for
calculations.}

\item{method}{the method used to calculate fold change. `method = "geometric"`
uses a log transform and a t-test to calculate fold change and estimate
confidence intervals. In the non-transformed space this is equivalent to
fold change using geometric means. `method = "median"` uses a log transform
and the method described in DOI: 10.1080/00949650212140 to calculate fold
change and estimate confidence intervals. In the non-transformed space this 
is equivalent to using group medians to calculate fold change.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Calculates fold change between groups for all features in a 
DatasetExperiment.
}
\examples{
D = MTBLS79_DatasetExperiment()
M = fold_change(factor_name='class')
M = model_apply(M,D)

}
