% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_int_class.R
\name{fold_change_int}
\alias{fold_change_int}
\title{Fold change for interactions between factors}
\usage{
fold_change_int(
  alpha = 0.05,
  factor_name,
  threshold = 2,
  control_group = character(0),
  method = "geometric",
  ...
)
}
\arguments{
\item{alpha}{(numeric) The p-value cutoff for determining significance. The default is \code{0.05}.}

\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{threshold}{(numeric) The fold change threshold for labelling features as significant. The default is \code{2}.}

\item{control_group}{(character) The level names of the groups used in the denominator (where possible) when computing fold change. One level for each factor, assumed to be in the same order as factor_name. The default is \code{character(0)}.}

\item{method}{(character) Fold change method. Allowed values are limited to the following: \itemize{\item{\code{"geometric"}: A log transform and a t-test is used to calculate fold change and estimate confidence intervals. In the non-transformed space this is equivalent to fold change using geometric means.}\item{\code{"median"}: A log transform and the method described by Price and Bonett to calculate fold change and estimate confidence intervals. In the non-transformed space this is equivalent to using group medians to calculate fold change.}} The default is \code{"geometric"}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{fold_change_int} object.
}
\description{
For more than one factor the fold change calculation is extended to include all combinations of levels (interactions) of all factors. Paired fold changes are not possible for this computation.
}
\examples{
D = MTBLS79_DatasetExperiment()
D=D[,1:10,drop=FALSE]
M = filter_smeta(mode='exclude',levels='QC',factor_name='class') +
    fold_change_int(factor_name=c('class','batch'))
M = model_apply(M,D)
}
\references{
Lloyd GR, Weber RJM (2021).
\emph{struct: Statistics in R Using Class-based Templates}.
R package version 1.3.1.
}
