% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_int_class.R
\name{fold_change_int}
\alias{fold_change_int}
\title{fold change for interactions class}
\usage{
fold_change_int(...)
}
\arguments{
\item{alpha}{confidence level to use for intervals}

\item{factor_name}{the sample_meta column to use}

\item{threshold}{a threshold to define fold change as 'significant'.}

\item{control_group}{a level of factor name to use as the control group for
calculations.}
}
\description{
Calculates fold change between groups for interactions between levels of
factors. Note that paired forced to FALSE for all comparisons.
}
\examples{
D = sbcms_DatasetExperiment()
D$data=D$data[,1:10,drop=FALSE]
M = filter_smeta(mode='exclude',levels='QC',factor_name='class') +
    fold_change_int(factor_name=c('class','batch'))
M = model_apply(M,D)

}
