% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_class.R
\name{fold_change_plot}
\alias{fold_change_plot}
\title{fold_change plot}
\usage{
fold_change_plot(number_features = 20, orientation = "portrait", ...)
}
\arguments{
\item{number_features}{The number of features to display on the plot}

\item{orientation}{The orientation of the plot (portrait or landscape). Portrait is default.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots fold change with error bars for a limited number of features.
}
\examples{
C = fold_change_plot()
}
