% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_selection_by_rank_class.R
\docType{class}
\name{forward_selection_byrank-class}
\alias{forward_selection_byrank-class}
\alias{forward_selection_byrank}
\title{forward selection by rank}
\description{
Forward selection by rank is a stepwise procedures that includes features
incrementally based on their rank. Any measure for ranking the features may
be used e.g. PLS VIP score, ttest p-value etc.
}
\examples{
M = forward_selection_byrank()
}
