% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_selection_by_rank_class.R
\docType{class}
\name{fs_line-class}
\alias{fs_line-class}
\alias{fs_line}
\title{forward_selection_plot}
\description{
plots the result of the evaluated models against the values the number of features within the search range.
}
