% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_class.R
\name{glog_opt_plot}
\alias{glog_opt_plot}
\title{glog transform optimisation plot}
\usage{
glog_opt_plot(plot_grid = 100, ...)
}
\arguments{
\item{plot_grid}{the resolution of the search space for plotting}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots the SSE error vs lambda for glog transform
}
\examples{
M = glog_opt_plot()
}
