% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_class.R
\name{glog_opt_plot}
\alias{glog_opt_plot}
\title{glog transform optimisation plot}
\usage{
glog_opt_plot(plot_grid = 100, ...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
plots the SSE error vs lambda for glog tranform
}
\examples{
M = glog_opt_plot()
}
