% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glog_class.R
\name{glog_transform}
\alias{glog_transform}
\title{glog transform}
\usage{
glog_transform(qc_label = "QC", factor_name, ...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
applies a glog transform to the input data
}
\examples{
M = glog_transform()
}
