% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_1d_class.R
\name{grid_search_1d}
\alias{grid_search_1d}
\title{grid_search_1d class}
\usage{
grid_search_1d(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
carries out a grid search for a single parameter
}
\examples{
M = grid_search_1d()
}
