% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_1d_class.R
\name{gs_line}
\alias{gs_line}
\title{Grid search line plot}
\usage{
gs_line(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
gs_line
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A plot of the calculated performance metric against the model input parameter values used to train the model. The optimum parameter value is indicated based on minimising (or maximising) the chosen metric.
}
\section{Inheritance}{

A \code{gs_line} object inherits the following \code{struct} classes: \cr\cr
\code{gs_line()} ⭢ \code{chart()} ⭢ \code{struct_class()}
}

\examples{
M = gs_line()

C = gs_line()
}
