% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_class.R
\name{HCA}
\alias{HCA}
\title{HCA method class}
\usage{
HCA(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
HCA method class. Calculate a hierarchical clustering for the input data
}
\examples{
M = HCA()
}
