% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_class.R
\name{hca_dendrogram}
\alias{hca_dendrogram}
\title{hca_dendrogram class}
\usage{
hca_dendrogram(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
plots a dendrogram for HCA
}
\examples{
C = hca_dendrogram()
}
