% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSD_class.R
\name{HSD}
\alias{HSD}
\title{HSD model class}
\usage{
HSD(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
HSD model class. Tukey's honest significant difference for ANOVA objects
}
\examples{
M = HSD()
}
