% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSDEM_class.R
\name{HSDEM}
\alias{HSDEM}
\title{HSD model class using estimated marginal means}
\usage{
HSDEM(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
HSD model class using estimate marginal means
}
\examples{
M = HSDEM()
}
