% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_xval_charts.R
\name{kfoldxcv_grid}
\alias{kfoldxcv_grid}
\title{kfoldxcv_grid class}
\usage{
kfoldxcv_grid(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
plot of cross validation results
}
\examples{
C = kfoldxcv_grid()
}
