% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_xval_charts.R
\name{kfoldxcv_grid}
\alias{kfoldxcv_grid}
\title{k-fold cross validation plot}
\usage{
kfoldxcv_grid(factor_name, level, ...)
}
\arguments{
\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{level}{(character) The level/group to plot.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{kfoldxcv_grid} object.
}
\description{
A graphic for visualising the true class and the predicted class of samples in all groups for all cross-validation folds.
}
\examples{
D = iris_DatasetExperiment()
I = kfold_xval(factor_name='Species') *
    (mean_centre() + PLSDA(factor_name='Species'))
I = run(I,D,balanced_accuracy())

C = kfoldxcv_grid(factor_name='Species',level='setosa')
chart_plot(C,I)

}
