% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_xval_charts.R
\name{kfoldxcv_metric}
\alias{kfoldxcv_metric}
\title{kfoldxcv_metric class}
\usage{
kfoldxcv_metric(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
box plot of cross validation results
}
\examples{
C = kfoldxcv_metric()
}
