% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_impute_class.R
\name{knn_impute}
\alias{knn_impute}
\title{knn missing value imputation}
\usage{
knn_impute(...)
}
\description{
applies a k-nearest neighbour approach to impute missing values
}
\examples{
M = knn_impute()
}
