% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_rank_sum_class.R
\name{kw_p_hist}
\alias{kw_p_hist}
\title{plot histogram of p values}
\usage{
kw_p_hist(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots a histogram of p values
}
\examples{
C = kw_p_hist()
}
