% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_rank_sum_class.R
\name{kw_p_hist}
\alias{kw_p_hist}
\title{Histogram of p values}
\usage{
kw_p_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
kw_p_hist
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A histogram of the p-values computed by the kruskal-wallis method
}
\examples{
C = kw_p_hist()
}
