% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_rank_sum_class.R
\name{kw_p_hist}
\alias{kw_p_hist}
\title{Histogram of p values}
\usage{
kw_p_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{kw_p_hist} object.
}
\description{
A histogram of the p-values computed by the kruskal-wallis method
}
\examples{
C = kw_p_hist()
}
