% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_rank_sum_class.R
\name{kw_rank_sum}
\alias{kw_rank_sum}
\title{kruskal-wallis model class}
\usage{
kw_rank_sum(...)
}
\description{
kruskal-wallis model class. Calculate kw-test for all features in a DatasetExperiment.
A non-parametric 1-way ANOVA.
}
\examples{
M = kw_rank_sum()

}
