% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_class.R
\name{linear_model}
\alias{linear_model}
\title{linear model class}
\usage{
linear_model(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
wrapper for R lm.
}
\examples{
M = linear_model()
}
