% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_class.R
\name{linear_model}
\alias{linear_model}
\title{linear model class}
\usage{
linear_model(formula, na_action = "na.omit", contrasts = list(), ...)
}
\arguments{
\item{formula}{The formula to use.}

\item{na_action}{The action to take when missing values are present. Can any
one of 'na.omit','na.fail','na.exclude' or 'na.pass'. Default is 'na.omit'.}

\item{contrasts}{The contrasts for this model. If zero length then the default contrasts are used.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Wrapper for R lm. See \code{lm} for details.
}
\examples{
M = linear_model()
}
