% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{log transform}
\usage{
log_transform(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
applies a log transform to the input data
}
\examples{
M = log_transform()
}
