% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{logarithm transform}
\usage{
log_transform(base = 10, ...)
}
\arguments{
\item{base}{(numeric) The base of the logarithm used for the transform. The default is \code{10}.\cr}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{log_transform} object with the following \code{output} slots:
\tabular{ll}{
\code{transformed} \tab          (DatasetExperiment) A DatasetExperiment object containing the log transformed data. \cr
}

struct object
}
\description{
A logarithmic transform is applied to all values in the data matrix.
}
\section{Inheritance}{

A \code{log_transform} object inherits the following \code{struct} classes: \cr\cr
\code{log_transform()} ⭢ \code{model()} ⭢ \code{struct_class()}
}

\examples{
M = log_transform(
      base = 10)

M = log_transform()
}
