% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_centre_class.R
\name{mean_centre}
\alias{mean_centre}
\title{mean_centre model class}
\usage{
mean_centre(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
Mean centres the columns a DatasetExperiment object
}
\examples{
M = mean_centre()
}
