% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_medians.R
\name{mean_of_medians}
\alias{mean_of_medians}
\title{Mean of median adjustment}
\usage{
mean_of_medians(factor_name, ...)
}
\arguments{
\item{factor_name}{the column sample of sample_meta to use. Mean of medians 
will be applied based on the levels in this factor.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Applies an offset to the data such that the mean of the medians is equal for all samples.
}
\examples{
D = iris_DatasetExperiment()
M = mean_of_medians(factor_name='Species')
M = model_apply(M,D)
}
