% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effect_class.R
\name{mixed_effect}
\alias{mixed_effect}
\title{Mixed Effects model class}
\usage{
mixed_effect(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
Mixed Effects model class. Applies RE model for all features in a DatasetExperiment
}
\examples{
M = mixed_effect()
}
