% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{mv_boxplot}
\alias{mv_boxplot}
\title{mv_boxplot class}
\usage{
mv_boxplot(
  label_outliers = TRUE,
  by_sample = TRUE,
  factor_name,
  show_counts = TRUE,
  ...
)
}
\arguments{
\item{label_outliers}{TRUE or FALSE to label outliers on the plot.}

\item{by_sample}{TRUE to plot missing values by sample, or FALSE to plot for features.}

\item{factor_name}{The sample_meta column to use.}

\item{show_counts}{TRUE to show a count of the number of items used to create the boxplot on the chart.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Boxplot of the numbers of missing values per sample/feature
}
\examples{
D = MTBLS79_DatasetExperiment()
C = mv_boxplot(factor_name='class')
chart_plot(C,D)

}
