% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_feature_filter_class.R
\name{mv_feature_filter}
\alias{mv_feature_filter}
\title{filter features by fraction of missing values}
\usage{
mv_feature_filter(
  threshold = 20,
  qc_label = "QC",
  method = "QC",
  factor_name,
  ...
)
}
\arguments{
\item{threshold}{The max percentage missing values in a feature, above which the feature is removed. Default  is 20.}

\item{qc_label}{The label of the QC samples in the named sample_meta column.}

\item{method}{"within_all" applies filter within classes,"within_one" applies
filter within any one class, "QC" applies filter within QC samples, "across" applies filter
ignoring class.}

\item{factor_name}{The name of the sample_meta column to use.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Filters features by the percent number of missing values, based on class labels if required.
}
\examples{
D = iris_DatasetExperiment()
M = mv_feature_filter(factor_name='Species',qc_label='versicolor')
M = model_apply(M,D)
}
