% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_feature_filter_class.R
\name{mv_feature_filter_hist}
\alias{mv_feature_filter_hist}
\title{plot for missing value sample filter}
\usage{
mv_feature_filter_hist(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
plots a histogram of % missing values per sample
}
\examples{
C = mv_feature_filter_hist()
}
