% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\docType{class}
\name{mv_histogram-class}
\alias{mv_histogram-class}
\alias{mv_histogram}
\title{mv_histogram class}
\description{
histograms indicating the numbers of missing values per sample/feature
}
