% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_chart_classes.R
\name{mv_histogram}
\alias{mv_histogram}
\title{mv_histogram class}
\usage{
mv_histogram(label_outliers = TRUE, by_sample = TRUE, ...)
}
\arguments{
\item{label_outliers}{[TRUE] or FALSE to label outliers on the plot}

\item{by_sample}{[TRUE] to plot by sample or FALSE to plot by features}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
histograms indicating the numbers of missing values per sample/feature
}
\examples{
D = MTBLS79_DatasetExperiment()
C = mv_histogram(label_outliers=FALSE,by_sample=FALSE)
chart_plot(C,D)

}
