% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_sample_filter_class.R
\name{mv_sample_filter_hist}
\alias{mv_sample_filter_hist}
\title{plot for missing value sample filter}
\usage{
mv_sample_filter_hist(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots a histogram of % missing values per sample
}
\examples{
C = mv_sample_filter_hist()
}
