% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_sample_filter_class.R
\name{mv_sample_filter_hist}
\alias{mv_sample_filter_hist}
\title{Histogram of missing values per sample}
\usage{
mv_sample_filter_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{mv_sample_filter_hist} object.
}
\description{
A histogram of the the proportion of missing values per sample
}
\examples{
C = mv_sample_filter_hist()
}
