% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nroot_transform_class.R
\name{nroot_transform}
\alias{nroot_transform}
\title{nroot transform}
\usage{
nroot_transform(root = 2, ...)
}
\arguments{
\item{root}{The nth root of the transform. Default is 2, resulting in a 
square-root transformation of the data.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Applies an nth root transform to the data
}
\examples{
M = nroot_transform()
}
