% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nroot_transform_class.R
\name{nroot_transform}
\alias{nroot_transform}
\title{nth root transform}
\usage{
nroot_transform(root = 2, ...)
}
\arguments{
\item{root}{(numeric) The nth root used for the transform. The default is \code{2}.\cr}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{nroot_transform} object with the following \code{output} slots:
\tabular{ll}{
\code{transformed} \tab          (DatasetExperiment) A DatasetExperiment object containing the nth root transformed data. \cr
}
}
\description{
All values in the data matrix are transformed by raising them to the power of 1/n.
}
\section{Inheritance}{

A \code{nroot_transform} object inherits the following \code{struct} classes: \cr\cr
\verb{[nroot_transform]} >> \verb{[model]} >> \verb{[struct_class]}
}

\examples{
M = nroot_transform(
      root = 2)

M = nroot_transform()
}
