% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_filter_class.R
\name{pairs_filter}
\alias{pairs_filter}
\title{Pairs filter}
\usage{
pairs_filter(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}

\item{factor_name}{(character) the column name of sample_meta containing the
labels}

\item{sample_id}{(character) the column name of sample_meta containing the
sample ids}
}
\value{
A STRUCT method object with functions for applying a pairs filter

struct object
}
\description{
Filters samples for paired analysis, ensuring each sample id is present in all groups.
}
\examples{
M=pairs_filter()

}
