% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_filter_class.R
\name{pairs_filter}
\alias{pairs_filter}
\title{Pairs filter}
\usage{
pairs_filter(factor_name, sample_id, ...)
}
\arguments{
\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{sample_id}{(character) Name of sample meta column containing sample identifiers.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{pairs_filter} object.

struct object
}
\description{
This filter is used for study designs with paired sampling to ensure that  measurements from the same source (e.g. patient) are represented in all factor levels and interactions.
}
\examples{
M=pairs_filter(factor_name='Class',sample_id='ids')
}
