% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoscale_class.R
\name{pareto_scale}
\alias{pareto_scale}
\title{Pareto scaling}
\usage{
pareto_scale(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
A STRUCT model object with methods for pareto scaling.

struct object
}
\description{
Pareto scaling centres the columns of the data in a DatasetExperiment object and divides by the square root of the standard deviation.
}
\examples{
D = iris_DatasetExperiment()
M = pareto_scale()
M = model_train(M,D)
M = model_predict(M,D)

}
