% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_class.R
\name{PCA}
\alias{PCA}
\title{PCA model class}
\usage{
PCA(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
Principal Component Analysis (PCA) model class. This object can be used to train/apply PCA mdoels to DatasetExperiment objects.
}
\examples{
M = PCA()
}
