% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_biplot_plot}
\alias{pca_biplot_plot}
\title{pca_biplot_plot class}
\usage{
pca_biplot_plot(...)
}
\description{
2d scatter plot of princpal component scores overlaid with principal component loadings.
}
\examples{
C = pca_biplot_plot()
}
