% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_correlation_plot}
\alias{pca_correlation_plot}
\title{PCA correlation plot}
\usage{
pca_correlation_plot(components = c(1, 2), ...)
}
\arguments{
\item{components}{(numeric) The Principal Components used to generate the plot. The default is \code{c(1, 2)}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
pca_correlation_plot
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A plot of the correlation between the variables/features and the selected principal component scores. Features with high correlation are well represented by the selected component(s)
}
\section{Inheritance}{

A \code{pca_correlation_plot} object inherits the following \code{struct} classes: \cr\cr
\verb{[pca_correlation_plot]} >> \verb{[chart]} >> \verb{[struct_class]}
}

\examples{
M = pca_correlation_plot(
      components = c(1, 2))

C = pca_correlation_plot()
}
