% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{PCA_dstat}
\alias{PCA_dstat}
\title{pca_dstat_plot class}
\usage{
PCA_dstat(number_components = 2, alpha = 0.05, ...)
}
\arguments{
\item{number_components}{The number of components to use.}

\item{alpha}{The confidence level to plot.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Bar chart showing mahalanobis distance from the mean in PCA scores space. A threshold is
plotted at a chosen confidence as an indicator for rejecting outliers.
}
\examples{
C = PCA_dstat()
}
