% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_loadings_plot}
\alias{pca_loadings_plot}
\title{pca_loadings_plot class}
\usage{
pca_loadings_plot(
  components = c(1, 2),
  style = "points",
  label_featurs = FALSE,
  ...
)
}
\arguments{
\item{components}{The principal components to plot (\code{numeric(2)})}

\item{style}{Plot style for loadings. Can be 'points' (default) or 'arrows'.}

\item{...}{additional slots and values passed to struct_class}

\item{label_features}{TRUE or FALSE to label features on the plot. Default is FALSE.}
}
\value{
struct object
}
\description{
2d scatter plot of princpal component loadings.
}
\examples{
C = pca_loadings_plot()
}
