% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_loadings_plot}
\alias{pca_loadings_plot}
\title{pca_loadings_plot class}
\usage{
pca_loadings_plot(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
2d scatter plot of princpal component loadings.
}
\examples{
C = pca_loadings_plot()
}
