% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_scores_plot}
\alias{pca_scores_plot}
\title{pca_scores_plot class}
\usage{
pca_scores_plot(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
2d scatter plot of princpal component scores.
}
\examples{
C = pca_scores_plot()
}
