% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plotfcns.R
\name{pca_scree}
\alias{pca_scree}
\title{pca_scree_plot class}
\usage{
pca_scree(...)
}
\description{
line plot showing percent variance and cumulative peercent variance for the computed components.
}
\examples{
C = pca_scree()
}
