% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\name{permutation_test}
\alias{permutation_test}
\title{permutation test class}
\usage{
permutation_test(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
Applies a permutation test to a model or model_seq()
}
\examples{
I=permutation_test()

}
