% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_class.R
\docType{class}
\name{permutation_test.scatter-class}
\alias{permutation_test.scatter-class}
\alias{permutation_test.scatter}
\title{permutation_test.scatter class}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
plots the results of a permutation test as histograms
}
\examples{
C = permutation_test.scatter()
}
