% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_sample_order_class.R
\name{permute_sample_order}
\alias{permute_sample_order}
\title{permute_sample_order class}
\usage{
permute_sample_order(number_of_permutations = 10, ...)
}
\arguments{
\item{number_of_permutations}{The number of times to permute the sample order}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Permutes the sample order a defined number of times, running the model each time
}
\examples{
C = permute_sample_order()
}
