% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\name{pls_regcoeff_plot}
\alias{pls_regcoeff_plot}
\title{pls_regcoeff_plot class}
\usage{
pls_regcoeff_plot(ycol = 1, ...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}

\item{level}{the group label to plot regression coefficients for}
}
\value{
struct object
}
\description{
Plots the regression coefficients of a PLSDA model.
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre()+PLSDA(factor_name='Species')
M = model_apply(M,D)

C = pls_regcoeff_plot(ycol='setosa')
chart_plot(C,M[2])
}
