% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_class.R
\name{PLSDA}
\alias{PLSDA}
\title{PLSDA model class}
\usage{
PLSDA(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
Partial least squares (PLS) discriminant analysis (DA) model class. This object can be used to train/apply PLS models.
}
\examples{
M = PLSDA()
}
