% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_class.R
\name{PLSDA}
\alias{PLSDA}
\title{Partial least squares discriminant analysis}
\usage{
PLSDA(number_components = 2, factor_name, ...)
}
\arguments{
\item{number_components}{(numeric, integer) The number of PLS components. The default is \code{2}.}

\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{PLSDA} object.
}
\description{
PLS is a multivariate regression technique that extracts latent variables maximising covariance between the input data and the response. The Discriminant Analysis variant uses group labels in the response variable and applies a threshold to the predicted values in order to predict group membership for new samples.
}
\details{
This object makes use of functionality from the following packages:\itemize{\item{\code{pls}}}
}
\examples{
M = PLSDA('number_components'=2,factor_name='Species')
}
\references{
Mevik B, Wehrens R, Liland K (2020).
\emph{pls: Partial Least Squares and Principal Component Regression}.
R package version 2.7-3, \url{https://CRAN.R-project.org/package=pls}.

Perez NF, Ferre J, Boque R (2009).
``Calculation of the reliability of classification in discriminant partial least-squares binary classification.''
\emph{Chemometrics and Intelligent Laboratory Systems}, \bold{95}(2), 122-128.

Barker M, Rayens W (2003).
``Partial least squares for discrimination.''
\emph{Journal of Chemometrics}, \bold{17}(3), 166-173.
}
