% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_class.R
\name{PLSDA}
\alias{PLSDA}
\title{Partial least squares discriminant analysis}
\usage{
PLSDA(number_components = 2, factor_name, ...)
}
\arguments{
\item{number_components}{(numeric, integer) The number of PLS components. The default is \code{2}.}

\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{PLSDA} object with the following \code{output} slots:
\tabular{ll}{
\code{scores} \tab          (data.frame)  \cr
\code{loadings} \tab          (data.frame)  \cr
\code{yhat} \tab          (data.frame)  \cr
\code{design_matrix} \tab          (data.frame)  \cr
\code{y} \tab          (data.frame)  \cr
\code{reg_coeff} \tab          (data.frame)  \cr
\code{probability} \tab          (data.frame)  \cr
\code{vip} \tab          (data.frame)  \cr
\code{pls_model} \tab          (list)  \cr
\code{pred} \tab          (data.frame)  \cr
\code{threshold} \tab          (numeric)  \cr
\code{sr} \tab          (data.frame) Selectivity ratio for a variable represents a measure of a variable's importance in the PLS model. The output data.frame contains a column of selectivity ratios, a column of p-values based on an F-distribution and a column indicating significance at p < 0.05. \cr
\code{sr_pvalue} \tab          (data.frame) A p-value computed from the Selectivity Ratio based on an F-distribution. \cr
}
}
\description{
PLS is a multivariate regression technique that extracts latent variables maximising covariance between the input data and the response. The Discriminant Analysis variant uses group labels in the response variable and applies a threshold to the predicted values in order to predict group membership for new samples.
}
\details{
This object makes use of functionality from the following packages:\itemize{\item{\code{pls}}}
}
\examples{
M = PLSDA('number_components'=2,factor_name='Species')
}
\references{
Liland K, Mevik B, Wehrens R (2021).
\emph{pls: Partial Least Squares and Principal Component Regression}.
R package version 2.8-0, \url{https://CRAN.R-project.org/package=pls}.

Perez NF, Ferre J, Boque R (2009).
``Calculation of the reliability of classification in discriminant partial least-squares binary classification.''
\emph{Chemometrics and Intelligent Laboratory Systems}, \bold{95}(2), 122-128.

Barker M, Rayens W (2003).
``Partial least squares for discrimination.''
\emph{Journal of Chemometrics}, \bold{17}(3), 166-173.
}
