% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_class.R
\name{PLSDA}
\alias{PLSDA}
\title{PLSDA model class}
\usage{
PLSDA(number_components = 2, factor_name, ...)
}
\arguments{
\item{number_components}{The number of PLS components to calculate.}

\item{factor_name}{The sample-meta column name to use.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Partial least squares (PLS) discriminant analysis (DA) model class. This object can be used to train/apply PLS models.
}
\examples{
M = PLSDA('number_components'=2,factor_name='Species')
}
