% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\name{plsda_predicted_plot}
\alias{plsda_predicted_plot}
\title{plsda_predicted_plot class}
\usage{
plsda_predicted_plot(factor_name, style = "boxplot", ...)
}
\arguments{
\item{factor_name}{The sample_meta column name to use}

\item{style}{The plot style. One of 'boxplot', 'violin' or 'density'.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
A box plot of the predicted values from a PLSDA model for each class. 
Only usitable for two class models.
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre()+PLSDA(factor_name='Species')
M = model_apply(M,D)

C = plsda_predicted_plot(factor_name='Species')
chart_plot(C,M[2])
}
