% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\name{plsda_roc_plot}
\alias{plsda_roc_plot}
\title{plsda_roc_plot class}
\usage{
plsda_roc_plot(factor_name, ...)
}
\arguments{
\item{factor_name}{The sample_meta column name to use}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the ROC curve of a PLSDA model. Only suitable for two classes.
}
\examples{
D = iris_DatasetExperiment()
M = mean_centre()+PLSDA(factor_name='Species')
M = model_apply(M,D)

C = plsda_roc_plot(factor_name='Species')
chart_plot(C,M[2])
}
