% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_charts.R
\name{plsda_scores_plot}
\alias{plsda_scores_plot}
\title{plsda_scores_plot class}
\usage{
plsda_scores_plot(...)
}
\description{
2d scatter plot of plsda component scores.
}
\examples{
C = plsda_scores_plot()
}
