% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSFC_class.R
\name{PLSFC}
\alias{PLSFC}
\title{PLS fold change}
\usage{
PLSFC(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
Partial least squares (PLS) fold change estimates
}
\examples{
C = PLSFC()
}
