% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{PLSR}
\alias{PLSR}
\title{PLSR model class}
\usage{
PLSR(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
Partial least squares (PLS) Regression model class. This object can be used to train/apply PLS models.
}
\examples{
M = PLSR()
}
