% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{PLSR}
\alias{PLSR}
\title{PLSR model class}
\usage{
PLSR(number_components = 2, factor_name, ...)
}
\arguments{
\item{number_components}{The number of PLS components to calculate.}

\item{factor_name}{The sample_meta column name to use.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Calculates a PLS regression model using the input data.
}
\examples{
M = PLSR()
}
