% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{PLSR}
\alias{PLSR}
\title{Partial least squares regression}
\usage{
PLSR(number_components = 2, factor_name, ...)
}
\arguments{
\item{number_components}{(numeric, integer) The number of PLS components. The default is \code{2}.}

\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{PLSR} object.
}
\description{
PLS is a multivariate regression technique that extracts latent variables maximising covariance between the input data and the response. For regression the response is a continuous variable.
}
\details{
This object makes use of functionality from the following packages:\itemize{\item{\code{pls}}}
}
\examples{
M = PLSR(factor_name='run_order')
}
\references{
Mevik B, Wehrens R, Liland K (2020).
\emph{pls: Partial Least Squares and Principal Component Regression}.
R package version 2.7-3, \url{https://CRAN.R-project.org/package=pls}.
}
