% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_cook_dist}
\alias{plsr_cook_dist}
\title{plsr_cook_dist class}
\usage{
plsr_cook_dist(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Cook's distance for a PLSR model. Cook's distance measures the effect of deleting
each observation. Samples with a larger Cook's distance might be considered outlying
as they have strong influence on the regression.
}
\examples{
C = plsr_cook_dist()
}
