% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_cook_dist}
\alias{plsr_cook_dist}
\title{Cook's distance barchart}
\usage{
plsr_cook_dist(ycol = 1, ...)
}
\arguments{
\item{ycol}{(numeric, integer, character) The y-block column to plot. The default is \code{1}.\cr}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
plsr_cook_dist
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A barchart of Cook's distance for each sample used to train a PLSR model. Cook's distance is used to estimate the influence of a sample on the model and can be used to identify potential outliers.
}
\section{Inheritance}{

A \code{plsr_cook_dist} object inherits the following \code{struct} classes: \cr\cr
\verb{[plsr_cook_dist]} >> \verb{[chart]} >> \verb{[struct_class]}
}

\examples{
M = plsr_cook_dist(
      ycol = 1)

C = plsr_cook_dist()
}
