% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_cook_dist}
\alias{plsr_cook_dist}
\title{Cook's distance barchart}
\usage{
plsr_cook_dist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{plsr_cook_dist} object.
}
\description{
A barchart of Cook㤼㸲s distance for each sample used to train a PLSR model. Cook's distance is used to estimate the influence of a sample on the model and can be used to identify potential outliers.
}
\examples{
C = plsr_cook_dist()
}
