% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_prediction_plot}
\alias{plsr_prediction_plot}
\title{PLSR prediction plot}
\usage{
plsr_prediction_plot(ycol = 1, ...)
}
\arguments{
\item{ycol}{(numeric, integer, character) The y-block column to plot. The default is \code{1}.\cr}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
plsr_prediction_plot
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A scatter plot of the true response values against the predicted values for a PLSR model.
}
\section{Inheritance}{

A \code{plsr_prediction_plot} object inherits the following \code{struct} classes: \cr\cr
\verb{[plsr_prediction_plot]} >> \verb{[chart]} >> \verb{[struct_class]}
}

\examples{
M = plsr_prediction_plot(
      ycol = 1)

C = plsr_prediction_plot()
}
