% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_prediction_plot}
\alias{plsr_prediction_plot}
\title{plsr_prediction_plot class}
\usage{
plsr_prediction_plot(...)
}
\description{
plots the true values against the predicted values for a PLSR model_
}
\examples{
C = plsr_prediction_plot()
}
