% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_qq_plot}
\alias{plsr_qq_plot}
\title{PLSR QQ plot}
\usage{
plsr_qq_plot(ycol = 1, ...)
}
\arguments{
\item{ycol}{(numeric, integer, character) The y-block column to plot. The default is \code{1}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{plsr_qq_plot} object.
}
\description{
A plot of the quantiles of the residuals from a PLSR model against the quantiles of a normal distribution.
}
\examples{
C = plsr_qq_plot()
}
