% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_qq_plot}
\alias{plsr_qq_plot}
\title{plsr_qq_plot class}
\usage{
plsr_qq_plot(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Quantiles of PLSR residuals against the qualtiles of a normal
distribution
}
\examples{
C = plsr_qq_plot()
}
