% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_qq_plot}
\alias{plsr_qq_plot}
\title{plsr_qq_plot class}
\usage{
plsr_qq_plot(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
plots the quantiles of PLSR residuals against the qualtiles of a normal
distribution
}
\examples{
C = plsr_qq_plot()
}
