% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_residual_hist}
\alias{plsr_residual_hist}
\title{PLSR residuals histogram}
\usage{
plsr_residual_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{plsr_residual_hist} object.
}
\description{
A histogram of the residuals for a PLSR model.
}
\examples{
C = plsr_residual_hist()
}
