% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_residual_hist}
\alias{plsr_residual_hist}
\title{plsr_residual_hist class}
\usage{
plsr_residual_hist(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
plots a histogram of the residuals
}
\examples{
C = plsr_residual_hist()
}
