% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSR_class.R
\name{plsr_residual_hist}
\alias{plsr_residual_hist}
\title{PLSR residuals histogram}
\usage{
plsr_residual_hist(ycol = 1, ...)
}
\arguments{
\item{ycol}{(numeric, integer, character) The y-block column to plot. The default is \code{1}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{plsr_residual_hist} object.
}
\description{
A histogram of the residuals for a PLSR model.
}
\examples{
C = plsr_residual_hist()
}
