% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqn_norm_method_class.R
\name{pqn_norm_hist}
\alias{pqn_norm_hist}
\title{PQN coefficient histogram}
\usage{
pqn_norm_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{pqn_norm_hist} object.
}
\description{
A histogram of the PQN coefficients for all features
}
\examples{
C = pqn_norm_hist()
}
