% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqn_norm_method_class.R
\name{pqn_norm_hist}
\alias{pqn_norm_hist}
\title{plot for PQN normalisation}
\usage{
pqn_norm_hist(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots a histogram of the PQN coeffients
}
\examples{
C = pqn_norm_hist()
}
