% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_na_class.R
\name{prop_na}
\alias{prop_na}
\title{prop_na model class}
\usage{
prop_na(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
prop_na Compares proportion of NA for all features in a DatasetExperiment
}
\examples{
M = prop_na()

}
