% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_class.R
\name{r_squared}
\alias{r_squared}
\title{Coefficient of determination (R-squared)}
\usage{
r_squared(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
r_squared
} object. This object has no \code{output} slots.
}
\description{
R-squared is a metric used to assess the goodness of fit for regression models. It measures how much variance of one variable can be explained by another variable.
}
\examples{
MET = r_squared()

}
