% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_filter.R
\name{rsd_filter}
\alias{rsd_filter}
\title{rsd filter}
\usage{
rsd_filter(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
filters features based on the relative standard deviation (RSD) for the QC samples
}
\examples{
M = rsd_filter()

}
