% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_filter.R
\name{rsd_filter_hist}
\alias{rsd_filter_hist}
\title{RSD histogram}
\usage{
rsd_filter_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{rsd_filter_hist} object.
}
\description{
A histogram of the calculated RSD values.
}
\examples{
C = rsd_filter_hist()

}
