% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd_filter.R
\name{rsd_filter_hist}
\alias{rsd_filter_hist}
\title{plot for rsd filter}
\usage{
rsd_filter_hist(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
plots a histogram of the calculated RSD for the RSD filter
}
\examples{
C = rsd_filter_hist()

}
