% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_corr.R
\name{sb_corr}
\alias{sb_corr}
\title{sbcms}
\usage{
sb_corr(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
Signal/batch correction using SMCBMS package
}
\examples{
M = sb_corr()
}
