% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbcms_dataset_class.R
\name{sbcms_dataset}
\alias{sbcms_dataset}
\title{Data from sbcms}
\usage{
sbcms_dataset()
}
\value{
dataset object
}
\description{
SBCMS package data after correction, as a struct dataset object
}
\examples{
D = sbcms_dataset()
summary(D)
}
