% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_classifier_class.R
\name{SVM}
\alias{SVM}
\title{SVM model classifier}
\usage{
SVM(
  factor_name,
  kernel = "linear",
  degree = 3,
  gamma = 1,
  coef0 = 0,
  cost = 1,
  class_weights = NULL,
  ...
)
}
\arguments{
\item{factor_name}{The sample-meta column name to use for group labels}

\item{kernel}{the kernel used in training and predicting. You
    might consider changing some of the following parameters, depending
    on the kernel type.\cr
    \describe{
      \item{linear:}{\eqn{u'v}{u'*v}}
      \item{polynomial:}{\eqn{(\gamma u'v + coef0)^{degree}}{(gamma*u'*v + coef0)^degree}}
      \item{radial basis:}{\eqn{e^(-\gamma |u-v|^2)}{exp(-gamma*|u-v|^2)}}
      \item{sigmoid:}{\eqn{tanh(\gamma u'v + coef0)}{tanh(gamma*u'*v + coef0)}}
      }
    }

\item{degree}{parameter needed for kernel of type \code{polynomial} (default: 3)}

\item{gamma}{parameter needed for all kernels except \code{linear}
    (default: 1/(data dimension))}

\item{coef0}{parameter needed for kernels of type \code{polynomial}
    and \code{sigmoid} (default: 0)}

\item{cost}{cost of constraints violation (default: 1)---it is the
    \sQuote{C}-constant of the regularization term in the Lagrange formulation.}

\item{class_weights}{a named vector of weights for the different classes, 
used for asymmetric class sizes. Not all factor levels have to be supplied 
(default weight: 1). All components have to be named. Specifying "inverse" 
will choose the weights inversely proportional to the class distribution.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Support vector machines model classifier. Wraps svm from the "e1071" package, which interfaces with the
"libsvm" library to train SVM classifiers.
}
\examples{
M = SVM(factor_name='Species',gamma=1)
}
