% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_classifier_class.R
\name{svm_plot_2d}
\alias{svm_plot_2d}
\title{SVM boundary plot (2d)}
\usage{
svm_plot_2d(factor_name, npoints = 100, ...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Plots the training data and the SVM boundary. 2d data only (ncol(D$data)==2).
}
\examples{
D = iris_DatasetExperiment()
M = SVM(factor_name='Species',kernel='linear')

C = svm_plot_2d(factor_name='Species')
chart_plot(C,M)

}
