% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_chart_class.R
\name{tic_chart}
\alias{tic_chart}
\title{Total Ion Count chart.}
\usage{
tic_chart(run_order, factor_name, ...)
}
\arguments{
\item{run_order}{(character) The column name of sample_meta indicating the run order of the samples.}

\item{factor_name}{(character) The name of a sample-meta column to use.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{tic_chart} object.
}
\description{
A scatter plot of Total Ion Count (sum of each sample) versus run order.
}
\examples{
D = iris_DatasetExperiment()
D$sample_meta$run_order=1:nrow(D)
C = tic_chart(factor_name='Species',run_order='run_order')
chart_plot(C,D)

}
