% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE_class.R
\name{tSNE}
\alias{tSNE}
\title{tSNE method class}
\usage{
tSNE(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
t-Distributed Stochastic Neighbor Embedding (tSNE) class. This object can be used to train/apply tSNE models to DatasetExperiment objects.
}
\examples{
M = tSNE()

}
