% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE_class.R
\name{tSNE_scatter}
\alias{tSNE_scatter}
\title{tSNE_scatter class}
\usage{
tSNE_scatter(factor_name, ...)
}
\arguments{
\item{factor_name}{Sample_meta column named used for colouring the points.}

\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
plots the new representation of data after applying tSNE
}
\examples{
M = tSNE_scatter(factor_name='Species')

}
