% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE_class.R
\name{tSNE_scatter}
\alias{tSNE_scatter}
\title{tSNE_scatter class}
\usage{
tSNE_scatter(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
plots the new representation of data after applying tSNE
}
\examples{
M = tSNE_scatter()

}
