% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{ttest}
\alias{ttest}
\title{t-test}
\usage{
ttest(
  alpha = 0.05,
  mtc = "fdr",
  factor_names,
  paired = FALSE,
  paired_factor = character(0),
  equal_variance = FALSE,
  ...
)
}
\arguments{
\item{alpha}{(numeric) The p-value cutoff for determining significance. The default is \code{0.05}.}

\item{mtc}{(character) Multiple test correction method. Allowed values are limited to the following: \itemize{\item{\code{"bonferroni"}: Bonferroni correction in which the p-values are multiplied by the number of comparisons.}\item{\code{"fdr"}: Benjamini and Hochberg False Discovery Rate correction.}\item{\code{"none"}: No correction.}} The default is \code{"fdr"}.}

\item{factor_names}{(character) The name of sample meta column(s) to use.}

\item{paired}{(logical) Apply a paired t-test. The default is \code{FALSE}.}

\item{paired_factor}{(character) The factor name that encodes the sample id for pairing. The default is \code{character(0)}.}

\item{equal_variance}{(logical) Equal variance. Allowed values are limited to the following: \itemize{\item{\code{"TRUE"}: The variance of each group is treated as being equal using the pooled variance to estimate the variance.}\item{\code{"FALSE"}: The variance of each group is not assumed to be equal and the Welch (or Satterthwaite) approximation is used.}} The default is \code{FALSE}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{ttest} object.
}
\description{
A t-test compares the means of two factor levels. Multiple-test corrected p-values are used to indicate the significance of the computed difference for all features.
}
\examples{
M = ttest(factor_name='Class')

}
