% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{ttest}
\alias{ttest}
\title{t-test model class}
\usage{
ttest(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
t-test model class. Calculate t-test for all features in a DatasetExperiment
}
\examples{
M = ttest()

}
