% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_norm_class.R
\name{vec_norm}
\alias{vec_norm}
\title{Vector normalisation}
\usage{
vec_norm(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{vec_norm} object.

struct object
}
\description{
The samples in the data matrix are normalised to account for differences in concentration by scaling each sample such that the sum of squares is equal to 1.
}
\examples{
M = vec_norm()

}
