% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_norm_class.R
\name{vec_norm}
\alias{vec_norm}
\title{Vector normalisation}
\usage{
vec_norm(...)
}
\arguments{
\item{...}{additional slots and values passed to struct_class}
}
\value{
struct object
}
\description{
Applies vector normalisation, such the sum of squared values for each sample
after normalisation are equal to 1.
}
\examples{
M = vec_norm()

}
