% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_norm_class.R
\name{vec_norm}
\alias{vec_norm}
\title{Vector normalisation}
\usage{
vec_norm(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{vec_norm} object with the following \code{output} slots:
\tabular{ll}{
\code{normalised} \tab          (DatasetExperiment) A DatasetExperiment object containing the normalised data. \cr
\code{coeff} \tab          (data.frame) The normalisation coefficients calculated by PQN. \cr
}

struct object
}
\description{
The samples in the data matrix are normalised to account for differences in concentration by scaling each sample such that the sum of squares is equal to 1.
}
\section{Inheritance}{

A \code{vec_norm} object inherits the following \code{struct} classes: \cr\cr
\code{vec_norm()} ⭢ \code{model()} ⭢ \code{struct_class()}
}

\examples{
M = vec_norm()

M = vec_norm()

}
