% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test_class.R
\name{wilcox_p_hist}
\alias{wilcox_p_hist}
\title{plot histogram of p values}
\usage{
wilcox_p_hist(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\description{
plots a histogram of p values
}
\examples{
M = wilcox_p_hist()

}
