% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test_class.R
\name{wilcox_p_hist}
\alias{wilcox_p_hist}
\title{Histogram of p values}
\usage{
wilcox_p_hist(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{
wilcox_p_hist
} object. This object has no \code{output} slots.
See \code{\link[struct]{chart_plot}} in the \code{struct} package to plot this chart object.
}
\description{
A histogram of p values for the wilcoxon signed rank test
}
\examples{
M = wilcox_p_hist()

}
