% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test_class.R
\name{wilcox_test}
\alias{wilcox_test}
\title{wilcoxon signed rank test}
\usage{
wilcox_test(
  alpha = 0.05,
  mtc = "fdr",
  factor_names,
  paired = FALSE,
  paired_factor = character(0),
  ...
)
}
\arguments{
\item{alpha}{(numeric) The p-value cutoff for determining significance. The default is \code{0.05}.}

\item{mtc}{(character) Multiple test correction method. Allowed values are limited to the following: \itemize{\item{\code{"bonferroni"}: Bonferroni correction in which the p-values are multiplied by the number of comparisons.}\item{\code{"fdr"}: Benjamini and Hochberg False Discovery Rate correction.}\item{\code{"none"}: No correction.}} The default is \code{"fdr"}.}

\item{factor_names}{(character) The name of a sample-meta column to use.}

\item{paired}{(logical) Apply a paired test. The default is \code{FALSE}.}

\item{paired_factor}{(character) The factor name containing sample ids for paired data. The default is \code{character(0)}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{wilcox_test} object.

struct object
}
\description{
A Mann-Whitney-Wilcoxon signed rank test compares ,the ranks of values in two groups. It is the non-parametric equivalent of a t-test. Multiple test corrected p-values are computed as indicators of significance for each variable/feature.
}
\examples{
M = wilcox_test(factor_name='class')

}
