% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test_class.R
\name{wilcox_test}
\alias{wilcox_test}
\title{wilcoxon signed rank test method class}
\usage{
wilcox_test(...)
}
\arguments{
\item{...}{slots and values for the new object}
}
\value{
struct object
}
\description{
wilcoxon signed rank test. Calculate signed rank test for all features in a
DatasetExperiment. Non-parametric ttest.
}
\examples{
M = wilcox_test()

}
